% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticOperators.R
\name{arithmeticCrossover}
\alias{arithmeticCrossover}
\title{Arithmetic Crossover}
\usage{
arithmeticCrossover(parent1, parent2, nFeatures)
}
\arguments{
\item{parent1, parent2}{vector of parent chromosomes in binary}

\item{nFeatures}{the number of features, which determines length of each vector}
}
\value{
a list of 2 child vectors produced from arithmetic crossover
}
\description{
Produces two new offsprings of same length given two parents using
single point crossover genetic operator
}
\details{
Linearly combines parents to produce two new offsprings according to
weighted recombination of vectors
}
