% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessFitness.R
\name{AssessFitness}
\alias{AssessFitness}
\title{Determine the fitness of some model}
\usage{
AssessFitness(individual, response, user.family = "gaussian", predictors,
  userfunc = FALSE)
}
\arguments{
\item{userfunc}{A fitness function that operates on a model, provided by the user. Defaults to FALSE. Built in options include "Residual" or "BIC"}
}
\description{
Called from within GeneticAlgorithmFit
}
\examples{
AssessFitness()
}
