% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCasignation.R
\name{AUCasignation}
\alias{AUCasignation}
\title{AUCasignation function}
\usage{
AUCasignation(population, yTrain, train)
}
\arguments{
\item{population}{bolean vector.}

\item{yTrain}{y values in train dataset.}

\item{train}{X values in train dataset.}
}
\description{
This function allows you to scores chromosomes.
}
\examples{
AUCasignation(population,yTrain,train)
}
\keyword{fitness}
\keyword{function}
