% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCNaiveBayes.R
\name{AUCNaiveBayes}
\alias{AUCNaiveBayes}
\title{AUCNaiveBayes function}
\usage{
AUCNaiveBayes(chromosome, yTrain, train, yTest, test)
}
\arguments{
\item{chromosome}{bolean vector.}

\item{yTrain}{y values in train dataset.}

\item{train}{X values in train dataset.}

\item{yTest}{y values in test dataset.}

\item{test}{X values in test dataset.}
}
\description{
This function allows you to scores chromosomes.
}
\examples{
AUCNaiveBayes(chromosome,yTrain,train,yTest,test)
}
\keyword{fitness}
\keyword{function}
