% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball}
\alias{baseball}
\title{Baseball Salaries Dataset}
\format{
A data frame with 337 rows and 28 variables:
\describe{
\item{salary}{Salary (in thousands of dollars)}
\item{average}{Batting average}
\item{obp}{On-base percentage}
\item{runs}{Number of runs}
\item{hits}{Number of hits}
\item{doubles}{Number of doubles}
\item{triples}{Number of triples}
\item{homeruns}{Number of home runs}
\item{rbis}{Number of runs batted in}
\item{walks}{Number of walks}
\item{sos}{Number of strike-outs}
\item{sbs}{Number of stolen bases}
\item{errors}{Number of errors}
\item{freeagent}{Indicator of "free agency eligibility"}
\item{arbitration}{Indicator of "arbitration eligibility"}
\item{runsperso}{runs/sos}
\item{hitsperso}{hits/sos}
\item{hrsperso}{homeruns/sos}
\item{rbisperso}{rbis/sos}
\item{walksperso}{walks/sos}
\item{obppererror}{obp/errors}
\item{runspererror}{runs/errors}
\item{hitspererror}{hits/errors}
\item{hrspererror}{homeruns/errors}
\item{soserrors}{sos\eqn{*}errors}
\item{sbsobp}{sbs\eqn{*}obp}
\item{sbsruns}{sbs\eqn{*}runs}
\item{sbshits}{sbs\eqn{*}hits}
}
}
\source{
\url{https://www.stat.colostate.edu/computationalstatistics/}
}
\usage{
baseball
}
\description{
A dataset containing offensive performance statistics such as batting average,
and number of home runs during the 1991 and 1992 seasons.
The player's 1992 salaries were also noted and serves as our response variable
in the testing of our genetic algorithm. These data are used in the book "Computational Statistics"
by G.H. Givens and J.A. Hoeting.
}
\keyword{datasets}
