% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{bitFlipMutation}
\alias{bitFlipMutation}
\title{Mutation}
\usage{
bitFlipMutation(child, rate, nFeatures)
}
\arguments{
\item{child}{vector of chromosomes in binary}

\item{rate}{numeric detailing rate of altering a feature}

\item{nFeatures}{the number of features, which determines length of each vector}
}
\value{
Mutated child with bits flipped with change  =  rate
}
\description{
Produces a new offspring of same length given a offspring using
mutation genetic operator
}
\details{
Altering each feature separately to produce a new
offspring given the rate and the child
}
