% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breed.R
\name{Breed}
\alias{Breed}
\title{Create an ensemble of 'children' from a given generation of parents}
\usage{
Breed(generation, fitness.vec, prob.mutate)
}
\arguments{
\item{generation}{A list of individual genomes or co-variate sets}

\item{fitness.vec}{A vector of the fitness values for the input \code{generation}}

\item{prob.mutate}{A natural rate of gene mutation determined in \code{\link{Select.R}}
or fixed by the user.}
}
\description{
\code{Breed()} takes a \code{generation} and its associated fitness data
  in \code{fitness.vec} as input and produces a number of children equal to the number
  of parents in \code{generation}.  This function allows for a natural rate of
  genetic mutation as set by \code{prob.mutate}.
}
\details{
\emph{Breed()} returns a list of 'genomes' for each child.

Called from within Select()
}
\examples{

\\code{\\link[GA]{CrossOverMutate}}
}
