% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentSelection.R
\name{calcFitness}
\alias{calcFitness}
\title{Calculates the fitness of a population}
\usage{
calcFitness(population, y, df, fitnessFUN, family)
}
\arguments{
\item{population}{matrix of population vectors}

\item{y}{vector of dependent variable in regression}

\item{df}{dataframe containing all covariates used for regression}

\item{fitnessFUN}{function handle for fitness function to be used}

\item{family}{a description of the error distribution
and link function to be used in the model.}
}
\value{
fitness values (AICs)
}
\description{
Calculates fitness of a population using AICs of regression results
}
