% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculate_fitness}
\alias{calculate_fitness}
\title{Calculate the fitness of a given model.}
\usage{
calculate_fitness(index, X, y, fit_func, family = gaussian)
}
\arguments{
\item{index}{A vector of indices of selected variables.}

\item{X}{Data frame containing all available features}

\item{y}{Response variable}

\item{fit_func}{Fitness function, with the default being AIC}

\item{family}{description of the family of distributions to be used.}
}
\value{
A fitness value of the provided model, based on the provided fitness function.
}
\description{
Calculate the fitness of a given model.
}
\examples{
fitness <- calculate_fitness(1:3, data.frame(replicate(10,sample(0:1,1000,rep=TRUE))), 1:1000, fit_func=AIC)
}
