% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{check_input}
\alias{check_input}
\title{Checking Modelling Input}
\usage{
check_input(input)
}
\arguments{
\item{input}{list of data.frames}
}
\description{
Checks whether the input is the appropriate format
}
\examples{
check_input(input = list(data.frame(a=1), data.frame(b=2)))
}
\author{
Louis Rémus
}
