% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosing.R
\name{chooseChromosomes}
\alias{chooseChromosomes}
\title{support function: chooseChromosomes()}
\usage{
chooseChromosomes(df_init, rankBased = TRUE)
}
\arguments{
\item{df_init}{is your initial set of chromosomes.}

\item{rankBased}{TRUE is choosing both parent based on their fitness and FALSE is choosing one parent based on fitness and the other one randomly. Defaults to TRUE.}
}
\description{
This function put choose/select the chromosomes for generating the next generation.
}
\examples{
chooseChromosomes()
}
\keyword{chooseChromosomes}
