% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chose_parents.R
\name{chose_parents_prop}
\alias{chose_parents_prop}
\title{chose_parents_prop}
\usage{
chose_parents_prop(individuals, objective = "AIC")
}
\arguments{
\item{individuals}{list of parents of type (variables, indices, linear_model)}

\item{objective}{objective function that defines the criterion to evaluate the
fitness of each individual}
}
\value{
the parent pairing - 2 parents samples with probabilities as defined in get_prob_individuals
}
\description{
Used to chose 2 parents proportionally to their fitness with respect
to the objective function
}
