% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R, R/chose_parents.R
\name{chose_parents_random}
\alias{chose_parents_random}
\alias{chose_parents_random}
\title{chose_parents_random}
\usage{
chose_parents_random(individuals)

chose_parents_random(individuals)
}
\arguments{
\item{individuals}{list of (variables, indices, linear_model)}

\item{individuals}{list of (variables, indices, linear_model)}
}
\value{
the parent pairing - 2 parents are randomly selected.

the parent pairing - 2 parents are randomly selected.
}
