% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fitness.R
\name{compute_fitness}
\alias{compute_fitness}
\title{Compute the fitness score}
\usage{
compute_fitness(Y, X, criteria, reg_model = lm)
}
\arguments{
\item{Y}{vector; target variable}

\item{X}{matrix/dataframe; design matrix (predictors)}

\item{criteria}{string; objective criteria to be used}
}
\value{
returns a vector of fitness level for each individual
}
\description{
This function calculates the fitness level
}
