% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{compute_population_goodness_of_fit}
\alias{compute_population_goodness_of_fit}
\title{Compute Population Goodness of Fit}
\usage{
compute_population_goodness_of_fit(data, population, regression_target,
  criterion = "AIC", verbose = FALSE)
}
\arguments{
\item{data}{data.frame: regression_target and covariates}

\item{population}{data.frame: selected column per element of the population}

\item{regression_target}{character: column of df regressed on other covariates in df}

\item{criterion}{character: goodness of fit criterion}
}
\value{
Returns the goodness of fit for the given population
}
\description{
Returns the goodness of fit for the given population
}
\author{
Louis Rémus, Jinhui Xu
}
