% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_next_generation.R
\name{create_next_generation}
\alias{create_next_generation}
\title{Create next generation from parent chromosomes.}
\usage{
create_next_generation(generation_t0, obj_fun_output, select_parents,
  crossover_method, crossover_parents, pCrossover, mutate_child, mutation_rate)
}
\arguments{
\item{generation_t0}{a matrix of parent chromosomes to be evaluated. Columns correspond to predictors/genes and rows correspond to parents/chromosomes.}

\item{obj_fun_output}{a numeric vector containg the objective function output for each parent chromosome.}

\item{select_parents}{function for selection of parents for \code{\link{crossover_parents}}. See \code{\link{select_parents}}.}

\item{crossover_method}{a character string indicating which crossover method to use: c("method1", "method2", "method3"). Default is "method1". See \code{\link{crossover_parents}}}

\item{crossover_parents}{a function for crossover between two parents. See \code{\link{crossover_parents}}.}

\item{pCrossover}{a number between 0 and 1 indicating the probability of crossover between two parents.}

\item{mutate_child}{a function for mutagenesis of child genes. See \code{\link{mutate_child}}.}

\item{mutation_rate}{a number between 0 and 1 indicating the probability of mutation. See \code{\link{mutate_child}}.}
}
\description{
Using fitness ranks of parent chromosomes, this function will select parents based upon their fitness rank, generate two children from each mate pair and mutate those child chromosomes. Function uses \code{\link{select_parents}}, \code{\link{crossover_parents}}, and \code{\link{mutate_child}} to generate children.
}
