% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{create_population}
\alias{create_population}
\title{create_population}
\usage{
create_population(chromosome_length, population_size)
}
\arguments{
\item{chromosome_length}{length of the chromosome of each individual in the population}

\item{population_size}{number of individuals in the population}
}
\value{
data frame with the population
}
\description{
Generates a population composed by individuals with chromosomes
}
\examples{
 create_population(chromosome_length = 5, population_size = 10)
 
}
