% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover of Chromosomes in Genetic Algorithm}
\usage{
crossover(pop, P_combn, P, C)
}
\arguments{
\item{pop}{a nested list represented the parent population of models or chromosomes}

\item{P_combn}{a list of all unique pairwise combinations of chromosome index (see \code{\link[utils]{combn}})}

\item{C}{a value representing chromosome size (number of 'alleles'). Should be of class integer, corresponding to the number of columns of input matrix x.}
}
\description{
Crossover is used in genetic algorithms for optimization when two parent chromosomes recombine to produce a child chromosome for the following generation.
Crossover randomly picks a location on the chrosome to split the genetic alleles of both parents and then recombines the segments of genes to produce a new chromosome.
}
