% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_crossover_mutation.R
\name{crossover_parents}
\alias{crossover_parents}
\title{Parent Chromosome Crossover}
\usage{
crossover_parents(generation_t0, parentInd, crossover_method, pCrossover,
  parent_rank)
}
\arguments{
\item{generation_t0}{a matrix of parent chromosomes to be evaluated. Columns correspond to predictors/genes and rows correspond to parents/chromosomes.}

\item{parentInd}{a vector containing the rank indexes of two parents selected for crossover.}

\item{crossover_method}{a character string indicating which method of crossover to be used, default is method 1, multipoint crossover.}

\item{pCrossover}{a number between 0 and 1 indicating the probability of crossover for each mate pair. Default is 0.8.}

\item{parent_rank}{a integer vector of fitness ranks for parent chromosomes.

References:
Geof H. Givens, Jennifer A. Hoeting (2013) Combinatorial Optimization (italicize). Chapter 3 of Computational Statistics (italicize).}
}
\description{
Parent chromosomes selected by \code{\link{select_parents}} will be used to generate new children using concept of crossover to combine genetic data to produce offspring. Parents are selected for crossover according to probability of crossover. If crossover does not occour, parent chromosomes are passed exactly to offpsring. This function produces two children from each parent chromosomes mate pair using one of three methods.
}
\details{
\describe{
\item{Method 1}{Multipoint crossover. Parent data is exchanged at three randomly selected crossover points along the chromosome to produce two offspring.}
\item{Method 2}{Probabilities for each gene are created from according to relative rankings of each parent and then child genotypes are sampled from \code{\link{rbinom}} using those probabilities.}
\item{Method 3}{Condordant mate pair genes are passed directly offspring, and non-corcordant genes are sampled from \code{\link{rbinom}} using probabilities proportional to relative ranking of each parent.}
}
}
