% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{currMutationRate}
\alias{currMutationRate}
\title{Computing Current Mutation Rate}
\usage{
currMutationRate(mutationRate, maxMutationRate, iterCounter, maxIter)
}
\arguments{
\item{mutationRate}{constant mutation rate. Value must be between 0 and 1.}

\item{maxMutationRate}{the mutation rate expected by the user when iteration limit is reached. If specified, then the mutation rate is linearly increasing in each iteration.}

\item{iterCounter}{an integer specifying which iteration in the  the current update}

\item{maxIter}{maximum number of iterations of updating the generation.}
}
\description{
\code{currMutationRate} computes the mutation rate in the current iteration of the genetic algorithm, assuming mutation rate is variable
}
