% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy Dataset for Testing}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
\item{x1}{Uniform Random Variable from 0 - 30}
\item{x2}{Uniform Random Variable from 0 - 30}
\item{x3}{Uniform Random Variable from 0 - 30}
\item{x4}{Uniform Random Variable from 0 - 30}
\item{x5}{Uniform Random Variable from 0 - 30}
\item{x6}{Uniform Random Variable from 0 - 30}
\item{x7}{Uniform Random Variable from 0 - 30}
\item{x8}{Uniform Random Variable from 0 - 30}
\item{x9}{Uniform Random Variable from 0 - 30}
\item{x10}{Uniform Random Variable from 0 - 30}
\item{y}{Linear combination of columns x1 - x5, with incercept and noise added}
}
}
\usage{
dummy_data
}
\description{
A dataset created for the purposes of testing the genetic algorithm correctly
identifies the features that contribute to the response variable. This dataset
contains 10 predictors, and 1 response variable that's a linear
combination of the 10 predictor, with an intercept and noise added.
The predictors are all Unif(0,30), the coefficients are norm(0,5), the
intercept is norm(100,10), the noise are norm(0,5), and all points are iid
generated from the seed 0.
}
\keyword{datasets}
