% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{support function: evaluation()}
\usage{
evaluation(X, Y, currentGeneration, popNum = 100, reg = "lm",
  criterion = "AIC")
}
\arguments{
\item{X}{matrix/dataframe; predictors}

\item{Y}{vector/matrix/dataframe; observations to be used in the fitting process}

\item{currentGeneration}{matrix; selectively produced generations}

\item{popNum}{integer; What is the population size of chromosomes? Default to 100}

\item{reg}{string; 'lm'/'glm; What is the regression type? Default to 'lm'}

\item{criterion}{string; function name; What is the objective criterion function? Default to 'AIC'}
}
\description{
This function evaluate the fitness value of chromosomes.
}
\examples{
evaluation()
}
\keyword{evaluation}
