% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractResponseVariable.R
\name{ExtractResponseVariable}
\alias{ExtractResponseVariable}
\title{Extract response variable from dataframe}
\usage{
ExtractResponseVariable(dataset, name)
}
\description{
This function allows the user to separate an input function
into reponse and predictor parts
}
\examples{
ExtractResponseVariable()
}
