% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{find_genes}
\alias{find_genes}
\title{find_genes}
\usage{
find_genes(chromosome, variables_names)
}
\arguments{
\item{chromosome}{chromosome to detect the active genes}

\item{variable_names}{name of the explanatory variables in the dataset}
}
\value{
vector with the active genes in each individual
}
\description{
Given a chromosome, it detects the active genes on it and returns the name of the gene on the original dataset
}
\examples{
 find_genes(chromosome = c(1,0,1), varaible_names = c("a","b","c"))
 
}
