% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations.R
\name{first_generation}
\alias{first_generation}
\title{first_generation}
\usage{
first_generation(y, dataset, population_size, interaction = F, most_sig = F,
  objective_function = "AIC", reg_method)
}
\arguments{
\item{y}{the response variable in the dataframe 'dataset'}

\item{dataset}{the data frame containing the variables in the model}

\item{population_size}{the number of individuals to generate}

\item{interaction}{a boolean indicating if we want to take into account
terms coming from interactions, False by default}

\item{most_sig}{a boolean indicating if we want to only consider the most
significant covariates}

\item{objective_function}{The objective criterion to use (default AIC).}

\item{reg_method}{regression method}
}
\value{
a list of individuals each a list with fields:
\item{variable}{the covariates kept in the linear model}
\item{indices}{the indices of these covariates in the data frame}
\item{linear_model}{the linear model with these covariates}
}
\description{
Performs the first iteration of the Genetic algorithm
Generates the first individuals
}
