% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_select.R
\name{fit_select}
\alias{fit_select}
\title{Selects fit individuals}
\usage{
fit_select(
  Y,
  X,
  models,
  fit_levs,
  f = 0.5,
  fselect = c("standard", "stochastic")
)
}
\arguments{
\item{Y}{vector; target variable}

\item{X}{matrix/dataframe; design matrix (predictors)}

\item{models}{matrix; current generation of individuals (candidate models)}

\item{fit_levs}{vector; fitness score for indivuals of current generation}

\item{f}{float; fraction of indiviuals to be selected from the current
generation}

\item{fselect}{string; selection criteria to be used}
}
\value{
returns a vector of fitness levels and variance of fitness in the
population
}
\description{
This function selects the best fitted individuals to be used as candidates
for next generation.
}
