% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Calculates fitness score}
\usage{
fitness(Y, X, models, core = 1, criteria = "AIC", reg_model = lm)
}
\arguments{
\item{Y}{vector; target variable}

\item{X}{matrix/dataframe; design matrix (predictors)}

\item{models}{matrix: individuals of the population (candidate models)}

\item{criteria}{string; objective criteria to be used}

\item{core:}{positive integer; No of cores to be used for parallization
(default 1 which means no parallization)}
}
\value{
returns a vector of fitness levels and variance of fitness
}
\description{
This function calculates the fitness level and the variance of fitness for all
individuals in the current generation and uses a helper function compute_fitn
ss to do that.
}
