% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitness.R
\name{fitness_aic}
\alias{fitness_aic}
\title{Fitness
Calculate the fitness of each individual in the population, the AICs() function works similarly but return AICs rather than fitness}
\usage{
fitness_aic(datax, datay, population, P, type)
}
\arguments{
\item{datax}{a n by p matrix, where n is the number of observations, and p is the total number of variables in the data}

\item{datay}{a length n numeric vector where n is the number of observations in the data}

\item{population}{a P by p matrix storing the population information, where P is the size of the population, and p is the gene length}

\item{P}{numeric, the size of population}

\item{type}{character, the type of model to use, 'lm' for linear regression, 'glm' for logistic regression}
}
\value{
Returns a numeric vector of individual fitnesses of the population
}
\description{
Fitness
Calculate the fitness of each individual in the population, the AICs() function works similarly but return AICs rather than fitness
}
\examples{
# create dummy data
x <- matrix(rnorm(100),10,10)
y <- rnorm(10)
P = 5
population = matrix(rep(c(1,0),25),5,10)
type = 'lm'

# run the fitness evaluation
fitness_aic(x,y,population,P,type)
}
\author{
Wei-Hsiang Sun
}
