% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{fitness_function}
\alias{fitness_function}
\title{fitness_function}
\usage{
fitness_function(formula, data, FUN = AIC, minimize = TRUE, ...)
}
\arguments{
\item{formula}{formula to fit to each individual}

\item{data}{dataset with all the variables of interest}

\item{FUN}{function to measure the fitness of an individual, default is AIC, it must take a glm() model as input}

\item{minimize}{depending on FUN, either minimize or not}

\item{...}{parameters for the glm model}
}
\value{
numeric fitness computed to the individual
}
\description{
Fitness function
}
\examples{
 data <-  matrix(runif(100), ncol=10, nrow = 10)
 fitness(formula = as.formula(y ~ x), data)
 
 Using the BIC function:
}
