% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_parallel}
\alias{fitness_parallel}
\title{Parallel Fitness Evaluation}
\usage{
fitness_parallel(pop, y, x, family, fitness_function, ncores)
}
\arguments{
\item{pop}{list of chromosomes to evaluate fitness of}

\item{y}{vector of responses}

\item{x}{matrix of covariates}

\item{family}{error distribution family for \code{\link[stats]{glm}}}

\item{fitness_function}{fitness function passed from \code{\link{select}}}

\item{ncores}{number of cores}
}
\description{
Evaluate fitness in parallel using \code{\link[parallel]{parLapply}} and return a numeric vector of results. Parallel evaluation
of chromosomes may be slower than calling serial \code{\link{fitness_serial}} unless P is very large.
}
