% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_serial}
\alias{fitness_serial}
\title{Serial Fitness Evaluation}
\usage{
fitness_serial(pop, y, x, family, fitness_function)
}
\arguments{
\item{pop}{list of chromosomes to evaluate fitness of}

\item{y}{vector of responses}

\item{x}{matrix of covariates}

\item{family}{error distribution family for \code{\link[stats]{glm}}}

\item{fitness_function}{fitness function passed from \code{\link{select}}}
}
\description{
Evaluate fitness in serial using \code{\link{vapply}} and return a numeric vector of results. Parallel fitness evaluation is provided through \code{\link{fitness_parallel}}
}
