% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessFunction.R
\name{FitnessFunction}
\alias{FitnessFunction}
\title{Determine the fitness of some model}
\usage{
FitnessFunction(model, userfunc = FALSE)
}
\arguments{
\item{model}{A model output from lm or glm}

\item{userfunc}{A fitness function that operates on a model, provided by the user. Defaults to FALSE. Built in options include "Residual" or "BIC"}
}
\description{
Called from within GeneticAlgorithmFit
}
\examples{
FitnessFunction()
}
