% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitnessRanks}
\alias{fitnessRanks}
\title{Genetic Algorithms Fitness Rank Operation}
\usage{
fitnessRanks(fitness)
}
\arguments{
\item{fitness}{The values of the fitness evaluated from the fitness
function.}
}
\value{
The ranked fitnesses.
}
\description{
Takes in fitness values and ranks them. See Page 80 of G/H.
}
\examples{
initPopAIC <- sapply(initPop, function(x) AIC(x$fit))
fitnessRanks(initPopAIC)

}
