\name{ga-class}
\docType{class}
\alias{ga-class}

\title{Class "ga"}

\description{An S4 class for genetic algorithms}

\section{Objects from the Class}{
  Objects can be created by calls to the \code{\link{ga}}.
}

\section{Slots}{
  \describe{
    \item{\code{call}}{}
    \item{\code{type}}{}
    \item{\code{min}}{}
    \item{\code{max}}{}
    \item{\code{nBits}}{}
    \item{\code{names}}{}
    \item{\code{popSize}}{}
    \item{\code{iter}}{}
    \item{\code{run}}{}
    \item{\code{maxiter}}{}
    \item{\code{suggestions}}{}
    \item{\code{population}}{}
    \item{\code{elitism}}{}
    \item{\code{pcrossover}}{}
    \item{\code{pmutation}}{}
    \item{\code{fitness}}{}
    \item{\code{best}}{}
    \item{\code{mean}}{}
    \item{\code{fitnessValue}}{}
    \item{\code{solution}}{}
  }
}

\author{Luca Scrucca}

\keyword{classes}

\seealso{\code{\link{ga}}}

%\examples{}
