\name{GA-package}
\alias{GA-package}
%\alias{ga-package}
\alias{GA}
\docType{package}

\title{Genetic Algorithms}

\description{
An R package for optimization using genetic algorithms. The package provides a flexible general-purpose set of tools for implementing genetic algorithms search in both the continuous and discrete case, whether constrained or not. Users can easily define their own objective function depending on the problem at hand. Several genetic operators are available and can be combined to explore the best settings for the current task. Furthermore, users can define new genetic operators and easily evaluate their performances. GAs can be run sequentially or in parallel.
}

%\details{
%GA: A Package for Genetic Algorithms in R \url{../doc/GA.pdf}
%Using parallel computing in GA package  \url{../doc/GAparallel.pdf}
%}

\seealso{The main function \code{\link{ga}}, and the \href{../doc/index.html}{documentation} accompanying the package.}

\author{Luca Scrucca \email{luca@stat.unipg.it}}

\section{References}{
Scrucca L (2013). GA: A Package for Genetic Algorithms in R.
\emph{Journal of Statistical Software}, 53(4), 1-37, \url{http://www.jstatsoft.org/v53/i04/}.
}

% \examples{}

\keyword{package}
