% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ga_compute}
\alias{ga_compute}
\title{GENETIC ALGORITHMS}
\usage{
ga_compute(dim, p, t = 100, selection_method = "rank",
  partial_update = FALSE, parent_ratio = 0.5, m_prob = 0.03, ...)
}
\arguments{
\item{dim}{A number, indicating the dimension of right-hand-side variables.}

\item{p}{A number, indicating the number of individuals in your population.}

\item{t}{A number, indicating the time of iterations.}

\item{selection_method}{A string, either 'rank' or 'score'. For 'rank', the algorithm will select parents according to their fitness ranks; for 'score', the algorithm will select parents according to their fitness raw scores}

\item{partial_update}{A logical variable, deciding whether to retain some parents of previous generation when updating population}

\item{parent_ratio}{A number between 0 and 1, deciding the ratio of parents retained from previous generation, if selection_method is TRUE}

\item{m_prob}{A number between 0 and 1, indicating the probability of mutation}

\item{...}{A set of parameters to decide how to calculate fitness function, explained below}

\item{data}{A data.frame of right-hand-variables, which should be (n by dim)}

\item{response}{A vector of response in your regression, which should be n by 1}

\item{fitness}{A function, to calculalte the fitness, default AIC}

\item{func}{A function of linear regression, either 'glm' or 'lm', default lm}

\item{min}{A logical variable, if TRUE, it will return the negative value of fitness scores}
}
\value{
A list of three :a list of latest generation, a vector of highest fitness score of each iteration, the best individual appeared(result & fitness score)
}
\description{
GENETIC ALGORITHMS
}
\examples{
GA_compute(dim = 500, p = 20, t = 100, selection_method = 'rank', partial_update = TRUE, data = data, fitness = AIC, func = glm, response = y)
}
