% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaMethod.R
\name{gaMethod}
\alias{gaMethod}
\title{Parent Selection}
\usage{
gaMethod(method, methodArg)
}
\arguments{
\item{population}{a matrix}

\item{fitnessVec}{a vector of length n or a matrix with n rows}

\item{population}{a matrix}
}
\value{
returns the selected parents
}
\description{
Parent Selection
}
\examples{

TBD
}
