\name{gaMonitor}
\alias{gaMonitor}

\title{Monitor genetic algorithm evolution}
\description{A function which prints the average and best fitness values at each iteration of GA search.
}
\usage{
gaMonitor(object, digits = getOption("digits"), \dots)
}

\arguments{
  \item{object}{an object of class \code{"ga"}, usually resulting from a call to function \code{\link{ga}}.}
  \item{digits}{minimal number of significant digits.}
  \item{\dots}{further arguments passed to or from other methods.}
}

%\details{}
\value{This function prints a summary of GA steps on the console.}

%\references{}

\author{Luca Scrucca}

\seealso{\code{\link{ga}}}

%\note{}

%\examples{}

%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
