% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{gather_fitness_generation}
\alias{gather_fitness_generation}
\title{gather_fitness_generation}
\usage{
gather_fitness_generation(generation, fitness_scores)
}
\arguments{
\item{generation}{dataframe containing the generation of individuals}

\item{fitness_scores}{estimated fitness per individual}
}
\value{
data frame with the information of the generation and its respective scores
}
\description{
Concatenate the generation with its respective fitness
}
\examples{
 generation <- as.data.frame(matrix(rbern(100), ncol=9, nrow = 10))
 scores <- runif(10)
}
