% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_controllable_data_gen.R
\name{generate_data}
\alias{generate_data}
\title{Generates normal data with 50 Covariates.}
\usage{
generate_data()
}
\value{
a 1000x51 matrix
}
\description{
This function creates a matrix that is 51 x 1000 where the first column is
 the response variables and the remain 50 are independent variables.
 The covariates are generated uniform(0,100)
 The parameter for each covariate is defined such that the
 first one is 20x0.9e1 followed by 20x0.9e2 with each subsequent
 parameter decaying until the last one is 20x0.9e50. This allows us to
 validate a genetic algorithm while generating a large dataset
}
\examples{

generate_data()


}
