% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_founders.R
\name{generate_founders}
\alias{generate_founders}
\title{Generate founders function}
\usage{
generate_founders(X, start_chrom)
}
\arguments{
\item{X}{a matrix or dataframe of predictor variables}

\item{start_chrom}{a numeric value indicating user specified number of founding chromosomes. Default is null.}
}
\description{
Generates founders given a covariate data matrix or dataframe and returns a matrix of founder chromosomes. Number of founders is determined by \code{\link{choose}(C, 2)} where C is number of predictors. Maximum number of founding chromosomes is 200. User can specify number of starting chromosomes and this number can be larger than 200. Founding chromosmes are generated randomly and unique values are kept.
}
