% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R, R/selection_mechanism.R
\name{generation_gap}
\alias{generation_gap}
\alias{generation_gap}
\title{Replace a proportion G of the old generation with their offspring.}
\usage{
generation_gap(old_gen, new_gen, G = nrow(new_gen))

generation_gap(old_gen, new_gen, G = nrow(new_gen))
}
\arguments{
\item{old_gen}{Output of ranked_models() of the previous generation.
A data frame.}

\item{new_gen}{Output of ranked_models() of the current generation. A data frame.}

\item{G}{Proportion of old_gen to replace with children.}

\item{children}{Output of ranked_models() of the children of the
previous generation. A data frame.}

\item{old_gen}{Output of ranked_models() of the previous generation. A data frame.}

\item{G}{Number of worst-performing offsprings the user wishes to replace. The default is the obeservation number of new_gen.}
}
\value{
The next generation, in descending order of fitness
  (most fit first). A data frame.

The updated new_gen.
}
\description{
Replace a proportion G of the old generation with their offspring.

Replace the worst-performing offsprings with their best-performing parents.
The function will automatically compare the G worst-performing offsprings with the G best-performing parents. Suppose among them, k parents outperform k offsprings; the parents will then replace the k offsprings.
}
\examples{
generation_gap(old_df, new_df, 10)
}
