% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationGAparallel.R
\name{generationGAparallel}
\alias{generationGAparallel}
\title{generationGAparallel function}
\usage{
generationGAparallel(population, yTrain, train, yTest, test, N, NChrSel,
  pMutation, pCrossover, paralell = TRUE)
}
\arguments{
\item{yTrain}{y values in train dataset.}

\item{train}{X values in train dataset.}

\item{yTest}{y values in test dataset.}

\item{test}{X values in test dataset.}

\item{N}{Initial population.}

\item{NChrSel}{Number of chromosomes to select.}

\item{pMutation}{Mutation probability.}

\item{pCrossover}{Crossover probability.}

\item{paralell}{Bolean indicationg if the process should run in parallel or sequential mode, defauls is paralell = TRUE.}

\item{chromosome}{bolean vector.}
}
\description{
This function allows you execute generations of GA.
}
\examples{
generationGAparallel(population,yTrain, train,yTest,test,N,NChrSel,pMutation,pCrossover,paralell = TRUE)
}
\keyword{Generation}
