% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticAlgorithmFit.R
\name{GeneticAlgorithmFit}
\alias{GeneticAlgorithmFit}
\title{Determine the fitness of some model}
\usage{
GeneticAlgorithmFit(dataset, response.name, userfunc = FALSE,
  user.family = "gaussian", flag.log.scale = TRUE, frac.replace = 0.2,
  Niter = 100, mutate.rate = FALSE, plot = TRUE)
}
\arguments{
\item{userfunc}{A fitness function that operates on a model, provided by the user. Defaults to FALSE. Built in options include "Residual" or "BIC"}

\item{model}{A model output from lm or glm}
}
\description{
This function takes a model object assesses its fitness
}
\examples{
GenericAlgorithmFit()
}
