% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_fitness}
\alias{get_fitness}
\title{get_fitness}
\usage{
get_fitness(data, name_y, generation, FUN = AIC, minimize = TRUE, ...)
}
\arguments{
\item{data}{dataset with the covariates and the variable the user is trying to estimate}

\item{name_y}{name of the variable that the user is trying to estimate}

\item{generation}{dataframe containing the generation of individuals}

\item{FUN}{function to measure the fitness of an individual, default is AIC, it must take a glm() model as input}

\item{minimize}{depending on FUN, either minimize or not}

\item{...}{parameters for the glm model
#' @return numeric vector with the fitness of the generation}
}
\description{
Compute the fitness of a generation
}
\examples{
 data <-  matrix(runif(100), ncol=10, nrow = 10)
 names(data) <- c("var1","var2","var3","var4","var5","var6","var7","var8","var9","var10")
 name_y <- "var8"
 generation <- matrix(rbern(100), ncol=9, nrow = 10)
 
}
