% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R,
%   R/get_fittest_individuals.R
\name{get_fittest_ind}
\alias{get_fittest_ind}
\alias{get_fittest_ind}
\alias{get_fittest_ind}
\title{Used to get the element in a list of elements like the result of
random_selection_regression}
\usage{
get_fittest_ind(individuals, objective = "AIC", fitness = F)

get_fittest_ind(individuals, objective = "AIC", fitness = F)

get_fittest_ind(individuals, objective = "AIC", fitness = F)
}
\arguments{
\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}

\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}

\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}
}
\value{
returns the element of individuals with that maximizes the objective function

returns the element of individuals with that maximizes the objective function

returns the element of individuals with that maximizes the objective function
}
\description{
Used to get the element in a list of elements like the result of
random_selection_regression

Used to get the element in a list of elements like the result of
random_selection_regression

Used to get the element in a list of elements like the result of
random_selection_regression
}
