% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_goodness_of_fit}
\alias{get_goodness_of_fit}
\title{Goodness of Fit}
\usage{
get_goodness_of_fit(df, regression_target, criterion = "AIC",
  glmindicator = FALSE, family = gaussian)
}
\arguments{
\item{df}{data.frame: regression_target and covariates}

\item{regression_target}{character: column of df regressed on other covariates in df}

\item{criterion}{character: goodness of fit criterion}
}
\value{
Goodness of fit value for the given regression problem
}
\description{
Returns the goodness-of-fit for regression_target ~ df
}
\author{
Louis Rémus, Jinhui Xu
}
