% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R,
%   R/get_fittest_individuals.R
\name{get_k_fittest_ind}
\alias{get_k_fittest_ind}
\alias{get_k_fittest_ind}
\alias{get_k_fittest_ind}
\title{Used to get the k elements from a list of (variables, indices, linear_model)
with the higgest objectives values}
\usage{
get_k_fittest_ind(individuals, objective = "AIC", fitness = F, k)

get_k_fittest_ind(individuals, objective = "AIC", fitness = F, k)

get_k_fittest_ind(individuals, objective = "AIC", fitness = F, k)
}
\arguments{
\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}

\item{k}{number of variables we want to choose.}

\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}

\item{individuals}{list with fields variables, indices, linear_model}

\item{objective}{name of the objective function to be maximized (as a string, inside quotes)}

\item{fitness}{only works with AIC, using a fitness function derived from the AIC}

\item{k}{number of variables we want to choose.}
}
\value{
returns the k elements of individuals with that have the highest objectives

returns the k elements of individuals with that have the highest objectives

returns the k elements of individuals with that have the highest objectives
}
\description{
Used to get the k elements from a list of (variables, indices, linear_model)
with the higgest objectives values

Used to get the k elements from a list of (variables, indices, linear_model)
with the higgest objectives values

Used to get the k elements from a list of (variables, indices, linear_model)
with the higgest objectives values
}
