% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_largest_interactions}
\alias{get_largest_interactions}
\title{get_largest_interactions}
\usage{
get_largest_interactions(y, dataset)
}
\arguments{
\item{y}{the response variable}

\item{dataset}{the dataframe containing the data}

\item{nb.var}{the threshold p-value, a covariate is kept if its p-value is
less than p_val}
}
\value{
a vector containing the names of the covariates that have been kept
}
\description{
Used to get the largest interaction terms.
}
