% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model}
\alias{get_model}
\title{get_model}
\usage{
get_model(candidate, method, X, ...)
}
\arguments{
\item{candidate}{(binary vector length c): on or off for each columns of X
method: method for fitting}

\item{X}{(matrix n x (c+1)): data (n x c) and the last column is the value of y.}
}
\value{
lm/glm object : the model selected after GA.
}
\description{
This function returns the parameter of the model once we fit method on candidate
}
