% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R
\name{get_most_significant_variables}
\alias{get_most_significant_variables}
\alias{get_most_significant_variables}
\title{Used to get the most significant covariates by recursively eliminating
the one with a p-value > p_val. This function can be used for the 
first generation in order to decrease the number of covariates that will
be taken into account in the iterations}
\usage{
get_most_significant_variables(dataset, y, p_val = 0.05)

get_most_significant_variables(dataset, y, p_val = 0.05)
}
\arguments{
\item{dataset}{the dataframe containing the data}

\item{y}{the response variable}

\item{p_val}{the threshold p-value, a covariate is kept if its p-value is
less than p_val}

\item{dataset}{the dataframe containing the data}

\item{y}{the response variable}

\item{p_val}{the threshold p-value, a covariate is kept if its p-value is
less than p_val}
}
\value{
a vector containing the names of the covariates that have been kept

a vector containing the names of the covariates that have been kept
}
\description{
Used to get the most significant covariates by recursively eliminating
the one with a p-value > p_val. This function can be used for the 
first generation in order to decrease the number of covariates that will
be taken into account in the iterations

Used to get the most significant covariates by recursively eliminating
the one with a p-value > p_val. This function can be used for the 
first generation in order to decrease the number of covariates that will
be taken into account in the iterations
}
