% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace population.R
\name{get_next_population}
\alias{get_next_population}
\title{Generate next population}
\usage{
get_next_population(parents, offsprings, scheme, pct = NULL)
}
\arguments{
\item{parents}{a data frame contains the original population}

\item{offsprings}{a data frame contains all offsprings}

\item{scheme}{"proportion" returns the first method and "re-rank" returns the second method}

\item{pct}{it is used to first method to determine the proportion of original population to be replaced}
}
\description{
get_next_population generate next population to use based on two methods.
The first method is to replace the worst certain percent of original population by offsprings.
The second method is to combine all parents and offsprings, and then choose the top individuals.
}
\author{
Xiao Li
}
