% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_calculation.R
\name{getFitness}
\alias{getFitness}
\title{Computing Fitness Scores}
\usage{
getFitness(objectiveVal)
}
\arguments{
\item{objectiveVal}{a vector of objective function values}
}
\description{
\code{getFitness} computes the fitness score of chromosomes based on their objective function values
}
