% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_calculation.R
\name{getObjective}
\alias{getObjective}
\title{Computing Objective Function Values}
\usage{
getObjective(X, y, pool, objectiveFunction, regressionType, nCores)
}
\arguments{
\item{X}{design matrix, in a dataframe or matrix}

\item{y}{a vector of responses}

\item{pool}{a matrix of booleans representing the chromosome pools}

\item{objectiveFunction}{accepts a function object to specify objective function. If provided by user, the function must be able to take glm object as input and return a numeric scalar.}

\item{regressionType}{a character string defining the distribution family.}

\item{nCores}{number of cores to use in evaluating the objective function value for each chromosome. If value is greater than 1, the evaluation is executed in parallel.}
}
\description{
\code{getObjective} returns a poolSize by 1 vector indicating the objective function value for each chromosome in the pool
}
