% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectparent.R
\name{getrankfitness}
\alias{getrankfitness}
\title{calculate the fitness based on the objective function's rank}
\usage{
getrankfitness(fitness, reverse = TRUE)
}
\arguments{
\item{fitness:}{goodness of fit ~ a casenum * 1 vector , default is AICvalue}

\item{reverse:}{whether we need to reverse the order when we rank the values in fitness, depends on whether the higher the value in the "fitness" vector , the better the fitness}
}
\description{
calculate the fitness based on the objective function's rank
}
\author{
Lei Zhang
}
