% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{init}
\alias{init}
\title{Initialization}
\usage{
init(p, datax)
}
\arguments{
\item{p}{numeric, the population size}

\item{datax}{a n by p matrix or dataframe, n is the number of observations and p is the number of variables in the original data}
}
\value{
a population matrix
}
\description{
Generate a random population as the first chromosome generation
}
\examples{
# create sample data
x <- matrix(100,10,10)
p <- 10

# initialize
init(p,x)
}
\author{
Wei-Hsiang Sun
}
