% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{init_parents}
\alias{init_parents}
\title{Initialization}
\usage{
init_parents(data, target, pop_size)
}
\arguments{
\item{data}{data.frame object corresponding to training set, dim = (N, 1 + predictor_size)}

\item{target}{y column indicated by name or index}

\item{pop_size}{the size of population}
}
\value{
Returns a pop_size by predictor_size matrix with columns named by predictor names
}
\description{
Generate a random population as the first chromosome generation
}
\examples{
example_df = data.frame(x1 = c(1,2,3,4,5), x2 = c(10,34,19,139,34), x3 = c(2,35,323,5,23), y = c(4,12,54,23,6))
init_parents(data = example_df, target = 'y', pop_size = 10)
}
\author{
Chloe Zhang
}
