% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pop.R
\name{init_pop}
\alias{init_pop}
\title{Initialize population}
\usage{
init_pop(N, X)
}
\arguments{
\item{N}{vector; Size of population}

\item{X}{matrix/dataframe; Number of predictors}
}
\value{
returns a matrix of initial candidate population
}
\description{
Initialize random population
}
