% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize}
\alias{initialize}
\title{Genetic Algorithms Initial Population Creation}
\usage{
initialize(Y, X, P, regType = "lm", family = "gaussian", seed = 1)
}
\arguments{
\item{Y}{The response vector as passed into select.}

\item{X}{The feature matrix as passed into select.}

\item{P}{The size of the population as passed or determined in select.}

\item{regType}{The type of regression, either 'lm' or 'glm'.}

\item{family}{The family for 'glm'.}

\item{seed}{The seed for reproducibility.}
}
\value{
A list of the P generated candidate solutions
where each entry of the list contains
\itemize{
\item \code{variables}: The variables that were
selected for the population.
\item \code{fit}: The \code{lm} or \code{glm} object of the fit with the
above variables.
}
}
\description{
Creates the initial population.
}
\examples{
Y <- mtcars$mpg
X <- mtcars[2:11]
P = 2 * ncol(X)
regType = 'lm'
family = 'gaussian'
seed = 1

initialize(Y, X, P, regType, family, seed)

}
