% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{initialize_parents}
\alias{initialize_parents}
\title{Generate the first generation of models.}
\usage{
initialize_parents(feature_count, generation_count = 2 * feature_count)
}
\arguments{
\item{feature_count}{Number of features given}

\item{P}{Number of parents in the generation}
}
\value{
A list of selected models, with each specified by the indices of selected columns.
}
\description{
Generate the first generation of models.
}
\examples{
parents <- initialize_parents(10, P = 10, number_features=4)
}
