% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initializeVectors}
\alias{initializeVectors}
\title{Initialize Population Vectors}
\usage{
initializeVectors(populationSize, nFeatures)
}
\arguments{
\item{populationSize}{size of initial population (# of vectors)}

\item{nFeatures}{the number of features, which determines length of each vector}
}
\value{
Matrix of size populationSize x nFeatures
denoting initialized population vectors
}
\description{
Randomly initialized n = populationSize vectors of length  =  nFeatures
}
