% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initiation.R
\name{initiation}
\alias{initiation}
\title{Initiation Function for Genetic Algorithm for Variable Selection}
\usage{
initiation(C, P)
}
\arguments{
\item{C}{The number of independent variables to be selected from}

\item{P}{Population size for each generation}
}
\description{
Generates initial parent generation to be used by select()
}
\examples{
# call initiation function
init_parents <- initiation( C = 5 , P = 30 )
}
\keyword{algorithm,}
\keyword{genetic}
\keyword{initiation}
\keyword{model}
\keyword{selection,}
