% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticOperators.R
\name{laplaceCrossover}
\alias{laplaceCrossover}
\title{Laplace Crossover - Source: Proceedings of the International Conference
on Soft Computing for Problem Solving [Deep and Thakur]}
\usage{
laplaceCrossover(parent1, parent2, nFeatures, locn = 0, scl = 0.35)
}
\arguments{
\item{parent1, parent2}{vector of parent chromosomes in binary}

\item{nFeatures}{the number of features, which determines length of each vector}

\item{locn}{the location parameter}

\item{scl}{the scaling parameter - most suitably 0.35 for integers}
}
\value{
a list of 2 child vectors produced from Laplace crossover
}
\description{
Produces two new offsprings of same length given two parents using
laplace crossover genetic operator
}
\details{
Generates child vectors based on Lapalce distribution within
the parent vectors
}
