% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R
\name{lit_sequence}
\alias{lit_sequence}
\title{Utility function to split a sequence in k subsequences
that is used in the tournament selection}
\usage{
lit_sequence(shuffled, k)
}
\arguments{
\item{shuffled}{1:n shuffled}

\item{k}{number of k subsequences}
}
\value{
a list of k subsequences
}
\description{
Utility function to split a sequence in k subsequences
that is used in the tournament selection
}
