% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main00.R
\name{main}
\alias{main}
\title{Main function}
\usage{
main(data, method, p, mutation_rate, regression_target, scheme)
}
\arguments{
\item{data}{data.frame: regression_target and covariates}

\item{method}{the method to select the parents, "twopropselection", "onepropselection" or "tournament"}

\item{p}{integer: splits number in crossover}

\item{mutation_rate}{numeric: the rate of mutation in population}

\item{regression_target}{character: Y variable}

\item{scheme}{method of replacing population, "proportion" or "re-rank"}
}
\value{
optimal x variable and every generation's AIC values
}
\description{
Returns the optimal x variable and every generation's AIC values
}
\author{
Jinhui Xu, Xiao Li
}
