% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{multipleCrossover}
\alias{multipleCrossover}
\title{Multiple Crossover}
\usage{
multipleCrossover(parents, numSplit)
}
\arguments{
\item{parents}{a 2 by C matrix of booleans, with each row representing a parent chromosome}

\item{numSplit}{number of split points in a crossover}
}
\description{
\code{multipleCrossover} accomplishes the process that parents crossover with multiple split points in their chromosomes. This function returns a vector of 2*C representing the genes of two children.
}
