% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate}
\alias{mutate}
\title{Genetic Algorithms Mutation Operation}
\usage{
mutate(rate, offspring)
}
\arguments{
\item{rate}{The rate of mutation.}

\item{offspring}{The candidate solution (a vector of 0's and 1's)
to potentially mutate.}
}
\value{
The mutated or unmutated candidate.
}
\description{
Performs a mutation operation for a given rate
on an offspring.
}
\examples{
Y <- mtcars$mpg
X <- mtcars[2:11]
P = 2 * ncol(X)
regType = 'lm'
family = 'gaussian'
seed = 1

initPop <- initialize(Y, X, P, regType, family, seed)
crossoverSingle <- crossover(initPop[[1]]$variables,
initPop[[2]]$variables, "single")

mutate(rate = .05, offspring = crossoverSingle$child1

}
