% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_crossover_mutation.R
\name{mutate_child}
\alias{mutate_child}
\title{Child Chromosome Mutagenesis.}
\usage{
mutate_child(mutation_rate, child, P, C)
}
\arguments{
\item{mutation_rate}{an optional numeric value between 0 and 1 indicating mutation rate.}

\item{child}{a vector containing the chromosome of child produced by crossover.}

\item{P}{an integer indicating the number of parent chromosomes.}

\item{C}{an integer indicating the number of predictor variables or genes.}
}
\description{
Child chromosomes produced using \code{\link{crossover_parents}} passed to this function will be modified using concept of genetic mutation, whereby a polymorphisms are produced according to a probably of mutatgenesis. Here, the default mutation rate is proportional to the size of the popluation, P, and number of chromosomes, C, \eqn{\frac{1}{P * C^{0.5}}}. User's can specify the mutation rate in \code{\link{select}}.
}
