% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation of Chromosomes in Genetic Algoritm
mutation randomly determines if and where on the chromosome, or model, a mutation of one of the genes, or covariates occurs. This is used in the genetic algorithm to introduce randomness not present in the genetic makeup of the population.}
\usage{
mutation(new_pop, C, C_ix, mutation)
}
\arguments{
\item{new_pop}{a nested list of chromosomes representing the child population. The the result of selection and crossover.}

\item{C}{a value representing the population size. Should be of class integer, corresponding to the number of columns of input matrix x for initial regression.}

\item{C_ix}{a sequence of integers from 1 to number of models considered for mutation (C).}

\item{mutation}{a value indicating the rate at which mutation should occur. Input as rate parameter lambda, or rpois for generating indices of chromosome list where mutation occurs.
Author(s):
Examples (?)}
}
\description{
mutation(new_pop,C,C_ix,mutation)
}
