% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor-selection.R
\name{Newpopulation}
\alias{Newpopulation}
\title{Survivor Selection}
\usage{
Newpopulation(offspringPop, offFitness, Pop.switch, replace.rate)
}
\arguments{
\item{offspringPop}{a matrix}

\item{offFitness}{a vector of length n}

\item{Pop.switch}{a matrix, where the first column is fitness, all other columns are genotypes}

\item{replace.rate}{a constant, proportion of offsprings to be replaced}

\item{c}{a constant specifies the exponential base in nonlinear ranking selection}

\item{k}{number of random selection from population in tournament selection}
}
\description{
Survivor Selection
}
\examples{
offspringPop <- matrix(rbinom(6*100, 1, prob = 0.5), 100, 6)
offFitness <- seq(15, 100, length.out=100)
replace.rate <- 0.1
Pop.switch <- cbind(seq(2, 50, length.out=10), matrix(rbinom(6*10, 1, prob = 0.5), 10, 6))
Newpopulation(offspringPop, offFitness, Pop.switch, replace.rate)

TBD
}
