\name{numericOrNA-class}
\docType{class}
\alias{numericOrNA-class}

\title{Virtual Class "numericOrNA" - Simple Class for subassignment Values}

\description{The class \code{"numericOrNA"} is a simple class union 
  (\code{\link{setClassUnion}}) of \code{"numeric"} and \code{"logical"}.
}

\section{Objects from the Class}{Since it is a virtual Class, no objects
  may be created from it.}

\examples{
showClass("numericOrNA")
}

\keyword{classes}
