% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{one_hot}
\alias{one_hot}
\title{one_hot}
\usage{
one_hot(p, n_var)
}
\arguments{
\item{p}{a list like the result of random_selection_regression}

\item{n_var}{the total number of indices}
}
\value{
a binary vector of length n_var with elements equal to 1 for the
indices in p$indices
}
\description{
utility function to encode indices in a binary vector
}
