% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectparent1.R
\name{onepropselection}
\alias{onepropselection}
\title{Method : onepropselection
Select one parent with probability proportional to fitness and to select
the other parent completely at random}
\usage{
onepropselection(originalparents, couplenum, fitness)
}
\arguments{
\item{originalparents}{: a casenum * variablenum matrix}

\item{couplenum}{: number of couples selected}

\item{fitness:}{goodness of fit ~ a casenum * 1 vector , default is AICvalue}
}
\description{
Returns the selected parents ~ list
}
\author{
Lei Zhang
}
