% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GA.R
\name{plot.GA}
\alias{plot.GA}
\title{Plotting function for GA object}
\usage{
\method{plot}{GA}(x, ..., col1 = "blue", col2 = "red")
}
\arguments{
\item{x}{a GA object as generated by \code{\link{select}}. Must be of class GA.}

\item{...}{a}

\item{col1}{a character string for color of chromosomes scatter plotpoints. Default is blue.}

\item{col2}{a character string for color of chromosomes mean and max lines. Default is red.}
}
\description{
Takes as input from \code{\link{select}} and creates a plot of fitness for all generations and lines plots of mean and max fitness across generations.
}
