% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_final.R
\name{plot_final}
\alias{plot_final}
\title{Plot goodness of fit vs generation}
\usage{
plot_final(input)
}
\arguments{
\item{input}{It is a list. The first element of the list contains all values of goodness of fit
during different generations.}
}
\description{
This function works to plot goodness of fit against generation
}
\author{
Xiao Li
}
