% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popInitial.R
\name{popInitial}
\alias{popInitial}
\title{support function: popInitial()}
\usage{
popInitial(popNum = 100, geneLength, zeroRate = 0)
}
\arguments{
\item{popNum}{What is the population size of chromosomes? Default to 100.}

\item{geneLength}{What is the gene length (positive integer)?}

\item{zeroRate}{What is the rate of zero among the genes of the first generation? Defaults to 0.}
}
\description{
This function intializes the population.
}
\examples{
popInitial()
}
\keyword{popInitial}
