% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{propotional}
\alias{propotional}
\title{Choose parents from generations propotional to their fitness}
\usage{
propotional(models, random = TRUE)
}
\arguments{
\item{models:}{dataframe with models (specified by the index of selected variables)
and their respective fitness value}

\item{random:}{if TURE, one parent will be selected randomly}
}
\value{
The selected parents(list of n lists, each list contains two parents)
}
\description{
Choose parents from generations propotional to their fitness
}
\examples{
x <- mtcars[-1]
y <- unlist(mtcars[1])
initial_index <- initialize_parents(10,20)$index
rankedmodels <- ranked_models(initial_index,x,y)
propotional(rankedmodels,random = T)
}
