% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_chromosome_swap.R
\name{random_chromosomes_swap}
\alias{random_chromosomes_swap}
\title{Function that executes alleles swap. It samples k indices in
range(number of genes) and swaps the alleles of the 2 parents
at these indices}
\usage{
random_chromosomes_swap(p1, p2, k, n_var)
}
\arguments{
\item{p1, p2}{the 2 parents: lists with fields (variables, indices, linear_model)}

\item{k}{number of locis to be swapped}

\item{n_var}{length of the chromosome}
}
\description{
Function that executes alleles swap. It samples k indices in
range(number of genes) and swaps the alleles of the 2 parents
at these indices
}
