% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_selection_regression.R
\name{random_selection_regression}
\alias{random_selection_regression}
\title{random_selection_regression}
\usage{
random_selection_regression(y, names, dataset, n, reg_method)
}
\arguments{
\item{y}{Name of the response variable in the dataset}

\item{names}{Names of the covariates in the dataset}

\item{dataset}{a dataset}

\item{n}{Number of covariates to include in the regression}

\item{reg_method}{Regression method}
}
\value{
A list with :
\item{variable}{the name of the covariates selected in the regression}
\item{indices}{their indices in the original vector names}
\item{linear_model}{the 'lm' object resulting from the regression}
}
\description{
select randomly the covariates to include in a regression and returns 
the names, indices of the covariates and an 'lm' object
}
