% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_objective_function.R
\name{rank_objective_function}
\alias{rank_objective_function}
\title{Rank parent chromosome fitness}
\usage{
rank_objective_function(obj_fun_output, minimize)
}
\arguments{
\item{obj_fun_output}{a numeric vector containg the objective function output for each chromosome for ranking.}

\item{minimize}{a logical value indicating whether to rank according to minimization or maximaziation optimization. Default is minimize.}
}
\description{
Parents are ranked according to their fitness. Rank order is inverse with east fit parent has a rank == 1.
}
