% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{ranked_models}
\alias{ranked_models}
\title{Fit the models and rank them by their fitness.}
\usage{
ranked_models(index, X, y, fit_func = AIC, family = gaussian,
  cluster = NA)
}
\arguments{
\item{index}{A list of indices of selected variables.}

\item{X}{Data frame of all features}

\item{y}{Dependent variable}

\item{fit_func}{Function for fitness measurement. Default is AIC.}

\item{family}{A description of the family of distributions to be used.}
}
\value{
a data frame containing index list and their respective fitness, sorted by fitness in ascending order
}
\description{
Fit the models and rank them by their fitness.
}
\examples{
X <- mtcars[-1]
y <- unlist(mtcars[1])
index <-initialize_parents(10,20)$index
ranked_models(index, X, y)
}
