% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_select.R
\name{rankSelect}
\alias{rankSelect}
\title{Selecting Parents by Rank}
\usage{
rankSelect(pool, fitness, oneParentRandom, numCrossoverSplit)
}
\arguments{
\item{pool}{a matrix of booleans representing the chromosome pools}

\item{fitness}{a vector indicating the fitness score of each chromosome in the pool}

\item{oneParentRandom}{a logical applying to rank-based selection. If TRUE, then we select one parent with the probability being its fitness score and the other randomly, otherwise, select both parents according to their fitness score.}

\item{numCrossoverSplit}{number of crossover points.}
}
\description{
\code{rankSelect} selects parents based on their fitness scores, which in turn depend on their rank in objective function values. In this operation, users have two choices in selecting parents as specified by the parameter "OneParenRandom". The default approach is to select each parent with probability equal to fitness, this approach is valid since the fitness score of all chromosome sums to one, indicating a selection probability. Another approach is to select a parent with probability proportional to fitness and to select the other parent completely at random.
}
