% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSelection.R
\name{rankSelection}
\alias{rankSelection}
\title{support function: rankSelection()}
\usage{
rankSelection(dfCurrent, usingRank = TRUE)
}
\arguments{
\item{dfCurrent}{is your initial set of chromosomes.}

\item{usingRank}{TRUE define selection probabilities using the rank and FALSE create selection probabilities proportionate to fitness. Defaults to TRUE.}
}
\description{
This function create selection probabilities of the chromosomes.
}
\examples{
rankSelection()
}
\keyword{rankSelection}
