% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regFunc}
\alias{regFunc}
\title{Genetic Algorithms Regression Operation}
\usage{
regFunc(regType, family, formula, data)
}
\arguments{
\item{regType}{The type of regression, either 'lm' or 'glm'.}

\item{family}{The family type for 'glm'.}

\item{formula}{The formula to evaluate.}

\item{data}{The data to evaluate the regression with.}
}
\value{
The regression object.
}
\description{
Allows for 'lm' or 'glm' to be passed in along with the
formula and data and performs the regression.
}
\examples{
regFunc('glm', 'gaussian', mpg ~ cyl, data = mtcars)

}
