% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{regression}
\usage{
regression(y, names, indices, dataset, reg_method)
}
\arguments{
\item{y}{Name of the response variable in the dataset}

\item{names}{Names of the covariates in the dataset}

\item{indices}{Indices of the covariates considered in the regression}

\item{dataset}{Name of the dataset}

\item{reg_method}{Regression method}
}
\value{
'lm' object obtained by performing the regression with the
 selected covariates
}
\description{
Performs an lm or glm given y, a list of covariates and the
indices of the covariates to take into account.
}
