% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_one_iteration}
\alias{run_one_iteration}
\title{Genetic Algorithm - One iteration}
\usage{
run_one_iteration(main_dataset, population, regression_target, criterion,
  couplenum, method, subsetnum, p, mutation_rate)
}
\arguments{
\item{main_dataset}{data.frame: regression_target and covariates}

\item{population}{data.frame: selected column per element of the population}

\item{regression_target}{character: column of df regressed on other covariates in df}

\item{criterion}{character: goodness of fit criterion}

\item{couplenum}{number of couples selected}

\item{method}{the method to select the parents, "twopropselection", "onepropselection" or "tournament"}

\item{subsetnum}{In method "tournament", we partition the set of chromosomes in generation t into "subsetnum" disjoint subsets}

\item{p}{Number of splits in the process of crossover.}
}
\value{
mutated_offspring
}
\description{
Performs one iteration of our genetic algorithm
}
\author{
Louis Rémus
}
