% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGA.R
\name{runGA}
\alias{runGA}
\title{runGA function}
\usage{
runGA(Nruns, N, NChrSel, pMutation, pCrossover, yTrain, train, yTest, test)
}
\arguments{
\item{Nruns}{Number of runs to execute.}

\item{N}{Initial population.}

\item{NChrSel}{Number of chromosomes to select.}

\item{pMutation}{Mutation probability.}

\item{pCrossover}{Crossover probability.}

\item{yTrain}{y values in train dataset.}

\item{train}{X values in train dataset.}

\item{yTest}{y values in test dataset.}

\item{test}{X values in test dataset.}
}
\description{
This function allows you execute a run of GA.
}
\examples{
require(GA)

Target variable for training
y = dataBalanced$train$target

Create training feature data frame
trainData = dataBalanced$train[,-1] 


resulTime = system.time(
  
 resultRunGA <- runGA(Nruns = 1,N = 2,NChrSel = 2,
                      pMutation = 0.05,pCrossover = 0.8,
                      yTrain = y,train = trainData,
                      yTest = y,test = trainData)
)


}
\keyword{cats}
