% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_crossover_mutation.R
\name{select_parents}
\alias{select_parents}
\title{Selection of Parents for Crossover}
\usage{
select_parents(parent_rank)
}
\arguments{
\item{parent_rank}{a vector indicating the rank of parents chromosomes. Ranking order is inverse: parent chromosomes with lowest fitness rank will have rank == 1.

References:
Geof H. Givens, Jennifer A. Hoeting (2013) Combinatorial Optimization (italicize). Chapter 3 of Computational Statistics (italicize).}
}
\description{
Selection of first parent is done according to fitness ranks generated with \code{\link{rank_objective_function}}. Probably of parent rank, \eqn{\frac{2\ r_{i}}{P(P+1)}}, from \eqn{i = 1,...,P}, where rank_i are fitness ranks of each parent chromosomes, and P is number of of parent chromosomes. The second parent is selected randomly and will not be the same as the previous parent. Fucntion returns ranks of select parents to be used for \code{\link{crossover_parents}}.
}
