% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Selection Of Models for Recombination}
\usage{
selection(pop_fitness, fitness, P, P_ix)
}
\arguments{
\item{pop_fitness}{a vector indicating fitness of each chrosomosome in the current generation based on the evaluated fitness function.}

\item{fitness}{a character indicating whether to use rank or weight when comparing model fitness.}

\item{P}{Population size, corresponding to the number of genes or covariates on each chromosome or model considered for recombination.}

\item{P_ix}{a sequence of integers, ending value P.

References:
Khalid Jebari, Mohammed Madiafi (2013) Selection Methods for Genetic Algorithms. https://www.researchgate.net/publication/259461147_Selection_Methods_for_Genetic_Algorithms.}
}
\description{
Standard selection is used to determine parent chromosomes fitness for recombination in the genetic algorithm for model optimization.
Fitness is assessed based on AIC or an inputted fitness function (must be function to minimize).
AIC is used in two ways: one through rank and another through weight.
Weight takes the absolute fitness to determine probability of recombination while rank uses relative ranking of fitness of models.
\itemize{
\item Selection of chromosomes that contribute their genes to generation t+1 done by resampling the population with probability proportional to weight or rank, with replacement.
}
}
