% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection_tournament}
\alias{selection_tournament}
\title{Tournament Selection Of Models for Recombination}
\usage{
selection_tournament(pop_fitness, fitness, P, P_ix, k)
}
\arguments{
\item{pop_fitness}{a vector indicating fitness of each chrosomosome in the current generation based on the evaluated fitness function.}

\item{fitness}{a character indicating whether to use rank or weight when comparing model fitness.}

\item{P}{Population size, corresponding to the number of genes or covariates on each chromosome or model considered for recombination.}

\item{P_ix}{a sequence of integers, ending value P.}

\item{k}{size of disjoin subset (must be smaller than P and whose quotient with P is 0, eg; P mod k = 0)

Author(s):
References:
Khalid Jebari, Mohammed Madiafi (2013) Selection Methods for Genetic Algorithms. https://www.researchgate.net/publication/259461147_Selection_Methods_for_Genetic_Algorithms.}
}
\description{
Standard selection is used to determine parent chromosomes fitness for recombination in the genetic algorithm for model optimization.
Fitness is assessed based on AIC or an inputted fitness function (must be function to minimize).
AIC is used in two ways: one through rank and another through weight.
\enumerate{
\item In tournament selection the population is first divded into a random partition of subsets of equal size k.
\item Then within each partition we select the most fit chromosome for propagation
}
}
