% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{set_formulas}
\alias{set_formulas}
\title{set_formulas}
\usage{
set_formulas(active_genes, name_y)
}
\arguments{
\item{active_genes}{character vector with the active genes in each individual}

\item{name_y}{string with the name of the variable the user is trying to estimate, target}
}
\value{
object type 'formula' per individual
}
\description{
Create the formulas to fit the model for each individual in the generation
}
\examples{
 set_formulas(active_genes = c("age","height","gender"), name_y = "weight")
 
}
