% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticOperators.R
\name{singlePtCrossover}
\alias{singlePtCrossover}
\title{Single Point Crossover}
\usage{
singlePtCrossover(parent1, parent2, nFeatures)
}
\arguments{
\item{parent1, parent2}{vector of parent chromosomes in binary}

\item{nFeatures}{the number of features, which determines length of each vector}
}
\value{
a list of 2 child vectors produced from uniform crossover
}
\description{
Produces two new offsprings of same length given two parents using
single point crossover genetic operator
}
\details{
Randomly choose an interger from 1 to nFeatures to be
the crossover point and swap the tails of its two parents to produce
two new offsprings
}
