% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R, R/chose_parents.R
\name{split_sequence}
\alias{split_sequence}
\alias{split_sequence}
\title{lit_sequence}
\usage{
split_sequence(shuffled, k)

split_sequence(shuffled, k)
}
\arguments{
\item{shuffled}{1:n shuffled}

\item{k}{number of k subsequences}

\item{shuffled}{1:n shuffled}

\item{k}{number of k subsequences}
}
\value{
a list of k subsequences

a list of k subsequences
}
\description{
Utility function to split a sequence in k subsequences
that is used in the tournament selection

Utility function to split a sequence in k subsequences
that is used in the tournament selection
}
