% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R, R/square2.R
\name{square}
\alias{square}
\title{Square}
\usage{
square(x)

square(x)
}
\arguments{
\item{x}{input value}
}
\value{
the square of x

the square of x
}
\description{
Gets the square of a number

Gets the square of a number
}
\examples{
square(3)

square(3)

}
