% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superEvaluation.R
\name{superEvaluation}
\alias{superEvaluation}
\title{support function: superEvaluation()}
\usage{
superEvaluation(X, Y, currentGeneration, popNum = 100, reg = "lm",
  criterion = "AIC")
}
\arguments{
\item{X}{is the first parent.}

\item{Y}{is the second parent.}

\item{currentGeneration}{is the initialized generation and the selectively produced generations.}

\item{popNum}{What is the population size of chromosomes? Default to 100.}

\item{reg}{What is the regression type? Default to lm.}

\item{criterion}{What is the objective criterion function? Default to AIC.}
}
\description{
This function is the parallelization version of evaluation().
}
\examples{
superEvaluation()
}
\keyword{superEvaluation}
