% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{sus}
\alias{sus}
\title{Proportional based selection}
\usage{
sus(fitness)
}
\arguments{
\item{fitness}{fitness for individual solutions}
}
\value{
Returns a matrix pop_size by 2 documenting the row index of each parent in population
}
\description{
Generate a Roulette Wheel but have two fixed points
Roll the wheel until it stops naturally
Select two parents with index where the two fixed points locate
Iterate this process for pop_size times
}
\examples{
sus(fitness = c(0.1, 0.14, 0.06, 0.1, 0.12, 0.24, 0.06, 0.03, 0.05, 0.1) )
}
\author{
Chloe Zhang
}
