% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper Function.R
\name{swapCol}
\alias{swapCol}
\title{swapCol}
\usage{
swapCol(dat, y_name)
}
\arguments{
\item{dat}{dataframe, data to generate offspring(default sets the dependent variable in first column and independent varialbes is other columns)}

\item{y_name}{string, column of dependent variable}
}
\value{
dataframe
}
\description{
swapCol
}
\details{
swaps the columns of the data set so that the first column corresponding to the dependent variable
}
