% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tournament}
\alias{tournament}
\title{Choose parents using tournament selection.
Each time, randomly pick k models from the previous generation and
the best-performed model will be selected as the parent.
Repeat this process until all slots for parents are filled.}
\usage{
tournament(models, k)
}
\arguments{
\item{models}{dataframe with models (specified by the index of selected variables)
and their respective fitness value. Object returned by ranked_models().}

\item{k}{size of each round of selection. Must be an integer smaller than generation size.}
}
\value{
The selected parents(list of ceiling(nrow(models)/2) lists, each list contains two parents)
}
\description{
Choose parents using tournament selection.
Each time, randomly pick k models from the previous generation and
the best-performed model will be selected as the parent.
Repeat this process until all slots for parents are filled.
}
\examples{
x <- mtcars[-1]
y <- unlist(mtcars[1])
initial_index <- initialize_parents(10,20)$index
rankedmodels <- ranked_models(initial_index,x,y)
tournament(rankedmodels,k = 2)
}
