% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R, R/chose_parents.R,
%   R/tournament_selection.R
\name{tournament_selection}
\alias{tournament_selection}
\alias{tournament_selection}
\alias{tournament_selection}
\alias{tournament_selection}
\title{A parents selection mechanism. The tournament selection splits the parents
population in k non overlapping subgroups and takes the fittest individual
in each group then takes n-k random subgroups (n the number of parents)
and takes the fittest individual in each of these random subgroups}
\usage{
tournament_selection(y, dataset, individuals, k = 4, n_var,
  objective = "AIC")

tournament_selection(y, dataset, individuals, k = 4, n_var,
  objective = "AIC")

tournament_selection(y, dataset, individuals, k = 4, n_var,
  objective = "AIC")

tournament_selection(y, dataset, individuals, k = 4, n_var,
  objective = "AIC")
}
\arguments{
\item{y}{response variable}

\item{dataset}{the dataframe containing the data}

\item{k}{the number of subsequences that should not overlap}

\item{n_var}{number of parents}

\item{objective}{the objective function}

\item{y}{response variable}

\item{dataset}{the dataframe containing the data}

\item{k}{the number of subsequences that should not overlap}

\item{n_var}{number of parents}

\item{objective}{the objective function}

\item{y}{response variable}

\item{dataset}{the dataframe containing the data}

\item{k}{the number of subsequences that should not overlap}

\item{n_var}{number of parents}

\item{objective}{the objective function}

\item{y}{response variable}

\item{dataset}{the dataframe containing the data}

\item{k}{the number of subsequences that should not overlap}

\item{n_var}{number of parents}

\item{objective}{the objective function}
}
\value{
a list of n_var parents that are the fittest ones in each subgroup

a list of n_var parents that are the fittest ones in each subgroup

a list of n_var parents that are the fittest ones in each subgroup

a list of n_var parents that are the fittest ones in each subgroup
}
\description{
A parents selection mechanism. The tournament selection splits the parents
population in k non overlapping subgroups and takes the fittest individual
in each group then takes n-k random subgroups (n the number of parents)
and takes the fittest individual in each of these random subgroups

A parents selection mechanism. The tournament selection splits the parents
population in k non overlapping subgroups and takes the fittest individual
in each group then takes n-k random subgroups (n the number of parents)
and takes the fittest individual in each of these random subgroups

A parents selection mechanism. The tournament selection splits the parents
population in k non overlapping subgroups and takes the fittest individual
in each group then takes n-k random subgroups (n the number of parents)
and takes the fittest individual in each of these random subgroups

A parents selection mechanism. The tournament selection splits the parents
population in k non overlapping subgroups and takes the fittest individual
in each group then takes n-k random subgroups (n the number of parents)
and takes the fittest individual in each of these random subgroups
}
