% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_select.R
\name{tournamentSelect}
\alias{tournamentSelect}
\title{Selecting Parents by Tournament}
\usage{
tournamentSelect(pool, fitness, numCrossoverSplit, groupNum)
}
\arguments{
\item{pool}{a matrix of booleans representing the chromosome pools}

\item{fitness}{a vector indicating the fitness score of each chromosome in the pool}

\item{numCrossoverSplit}{number of crossover points.}

\item{groupNum}{number of groups to be partitioned in each round of the tournament selection. Value must be between 2 and poolSize / 2 (inclusive) to ensure meaningful tournaments}
}
\description{
\code{tournamentSelect} selects parents by tournaments. The chromosomes in the parent generation is randomly paritioned into k disjoint subsets of equal size and perform a tournament amongst them. The individual with best objective function (e.g. least AIC score) in each group is chosen as a parent for the next generation. The function takes "groupNum" as one of the parameters, user may use the default value for which the "groupNum' is set to be the value that ensure each group size equals to 2, or user may define the number of subset they'd like to obtain. This process is repeated until the desired amount of population has been reached. Parents are then paired randomly for breeding.
}
