% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{training}
\alias{training}
\title{training}
\usage{
training(candidate, method, X, fitness_function, ...)
}
\arguments{
\item{candidate}{(binary vector length c): on or off for each columns of X}

\item{X}{(matrix n x (c+1)): data (n x c) and the last column is the value of y.}

\item{method:}{method for fitting lm/glm}

\item{fitness_function:}{error of the model}
}
\value{
fitness_value (float): fitness value of the model
}
\description{
This function fits the method on candidates and return the fitness value of the candidate
}
