% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{twopropselection}
\alias{twopropselection}
\title{Proportional Selection based}
\usage{
twopropselection(fitness)
}
\arguments{
\item{fitness}{fitness for individual solutions from fit}
}
\value{
Returns a matrix pop_size by 2 documenting the row index of each parent in population
}
\description{
Select one parent with probability proportional to fitness
Select the other parent randomly
Iterate this process for pop_size times
}
\examples{
twopropselection(fitness = c(0.1, 0.14, 0.06, 0.1, 0.12, 0.24, 0.06, 0.03, 0.05, 0.1))
}
\author{
Chloe Zhang
}
