% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticOperators.R
\name{uniformCrossover}
\alias{uniformCrossover}
\title{Uniform Crossover}
\usage{
uniformCrossover(parent1, parent2, nFeatures)
}
\arguments{
\item{parent1, parent2}{vector of parent chromosomes in binary}

\item{nFeatures}{the number of features, which determines length of each vector}

\item{rate}{numeric detailing rate of obtaining parent1's traits}
}
\value{
a list of 2 child vectors produced from uniform crossover
}
\description{
Produces two new offspring of same length given two parents using
uniform crossover genetic operator
}
\details{
Altering each feature separately by the rate to
produce a new offspring for each parent
}
