% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R, R/iterations.R
\name{update_generations}
\alias{update_generations}
\alias{update_generations}
\title{Used to perform an iteration of the Genetic Algorithm and return the
next generation from the generation 'individuals'}
\usage{
update_generations(y, dataset, individuals, objective, pop_size, generation_gap,
  parent_selection, nb_groups, gene_selection, gene_operator, nb_pts,
  reg_method, mu)

update_generations(y, dataset, individuals, objective, pop_size, generation_gap,
  parent_selection, nb_groups, gene_selection, gene_operator, nb_pts,
  reg_method, mu)
}
\arguments{
\item{y}{the name of the response variable in the data frame 'dataset'}

\item{dataset}{the data frame containing the variables in the model}

\item{pop_size}{the number of individuals in the next generation}

\item{generation_gap}{the proportion of individuals to bereplaced
generation to keep in the next one}

\item{nb_groups}{the number of subgroups if the tournament selection
is the parent selection mechanism}

\item{gene_selection}{the gene operator, chose between "crossover",
"random" for random locis swap or provide the name of your own function
inside quotation marks}

\item{selection}{the name of the parents selection mechanism, the user
can provide the name of his own function defining a parent selection
mechanism}

\item{nb_points}{the number of crossover points if gene_selection="crossover"}

\item{y}{the name of the response variable in the data frame 'dataset'}

\item{dataset}{the data frame containing the variables in the model}

\item{pop_size}{the number of individuals in the next generation}

\item{generation_gap}{the proportion of individuals to bereplaced
generation to keep in the next one}

\item{selection}{the name of the parents selection mechanism, the user
can provide the name of his own function defining a parent selection
mechanism}

\item{nb_groups}{the number of subgroups if the tournament selection
is the parent selection mechanism}

\item{gene_selection}{the gene operator, chose between "crossover",
"random" for random locis swap or provide the name of your own function
inside quotation marks}

\item{nb_points}{the number of crossover points if gene_selection="crossover"}
}
\value{
a list of pop_size new models each a list with fields
\item{variables}{the covariates kept in the model}
\item{indices}{the indices of the covariates kept in the model}
\item{linear_model}{the linear model for these covariates}

a list of pop_size new models each a list with fields
\item{variables}{the covariates kept in the model}
\item{indices}{the indices of the covariates kept in the model}
\item{linear_model}{the linear model for these covariates}
}
\description{
Used to perform an iteration of the Genetic Algorithm and return the
next generation from the generation 'individuals'

Used to perform an iteration of the Genetic Algorithm and return the
next generation from the generation 'individuals'
}
