% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pool.R
\name{updatePool}
\alias{updatePool}
\title{Updating Chromosome Pool}
\usage{
updatePool(
  pool,
  fitness,
  tournamentSelection,
  groupNum,
  oneParentRandom,
  numCrossoverSplit,
  mutationRate,
  maxMutationRate,
  iterCounter,
  maxIter
)
}
\arguments{
\item{pool}{a matrix of booleans representing the chromosome pools}

\item{fitness}{a vector indicating the fitness score of each chromosome in the pool}

\item{tournamentSelection}{a logical indicating whether to use tournament selection. If FALSE, rank-based selection is used.}

\item{groupNum}{number of groups to be partitioned in each round of the tournament selection. Value must be between 2 and poolSize / 2 (inclusive) to ensure meaningful tournaments}

\item{oneParentRandom}{a logical applying to rank-based selection. If TRUE, then we select one parent with the probability being its fitness score and the other randomly, otherwise, select both parents according to their fitness score.}

\item{numCrossoverSplit}{number of crossover points.}

\item{mutationRate}{constant mutation rate. Value must be between 0 and 1.}

\item{maxMutationRate}{the mutation rate expected by the user when iteration limit is reached. If specified, then the mutation rate is linearly increasing in each iteration}

\item{iterCounter}{an integer specifying which iteration in the  the current update}

\item{maxIter}{maximum number of iterations of updating the generation.}
}
\description{
\code{updatePool} is used to update the chromosome pool in each iteration of the genetic algorithms. It returns a vector of length P*C, which is loaded by all genes of next generation chromosomes.
}
