% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed.R
\name{breed}
\alias{breed}
\title{Breed}
\usage{
breed(
  current_gen,
  pop_size,
  crossover_count = 1,
  mut_prob = 0.01,
  mutations = 5
)
}
\arguments{
\item{current_gen}{matrix; chromosomes for the current generation}

\item{pop_size}{positive integer; size of the original population (no of individuals)}

\item{crossover_count}{vector; number of genes to be cross overed (default set
as 1)}

\item{mut_prob}{float; probability of mutation (between 0 and 1) (default is
.05)}

\item{mutations}{positive integer; number of genes to be mutated if selected
to mutate (default  set to 1)}
}
\value{
returns a matrix of new generation
}
\description{
The function breed a new population given the current population by doing
cross-over and mutation of the individuals from current population. It uses
two helper functions, cross over and mutation. Two candidate parents
are selected everytime at random from the current population to be cross
overed and mutated to create offsprings for the new generation. Function makes
sure to breed the same number of individuals in the new generation as the
original population size.
}
