% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breed.R
\name{Breed}
\alias{Breed}
\title{Determine the fitness of some model}
\usage{
Breed(generation, fitness.vec, predictors, prob.mutate)
}
\arguments{
\item{model}{A model output from lm or glm}

\item{userfunc}{A fitness function that operates on a model, provided by the user. Defaults to FALSE. Built in options include "Residual" or "BIC"}
}
\description{
This function takes a model object assesses its fitness
}
\examples{
Breed()
}
