% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chose_parents.R
\name{chose_parents_prop_random}
\alias{chose_parents_prop_random}
\title{chose_parents_prop_random}
\usage{
chose_parents_prop_random(individuals, objective = "AIC")
}
\arguments{
\item{individuals}{list of (variables, indices, linear_model) 
see random_selection_regression}

\item{objective}{objective function}
}
\value{
the parent pairing - 1 parent samples with probabilities and 1 parent is randomly selected.
}
\description{
Used to chose one parent at random and the other proportionally to
their fitness with regard to their fitness function
}
