% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossOver.R
\name{crossOver}
\alias{crossOver}
\title{crossOver function}
\usage{
crossOver(chromosome, pCrossover, chromosomes)
}
\arguments{
\item{chromosome}{bolean vector.}

\item{pCrossover}{Crossover probability.}

\item{chromosomes}{Matrix with chromosomes. Each column is chromosomes and rows the alleles.}
}
\description{
This function allows you to generate chromosomes by subsamples of parents.
}
\examples{
crossOver(chromosome,pCrossover,chromosomes)
}
\keyword{Mix}
\keyword{chromosomes.}
