% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossover}
\alias{crossover}
\title{Genetic Algorithms Crossover Operation}
\usage{
crossover(parent1, parent2, type = "single", num_splits = NA)
}
\arguments{
\item{parent1}{The first candidate solution (a vector of 0's and 1's).}

\item{parent2}{The second candidate solution (a vector of 0's and 1's).}

\item{type}{The type of crossover, either 'single' or 'multiple'.}

\item{num_splits}{The number of splits for type multiple.
If invalid \code{num_splits} specified, then will default to 2.}
}
\value{
A list of vectors of the children to return.
}
\description{
Performs a crossover operation for the specified type and
parents.
}
\examples{
Y <- mtcars$mpg
X <- mtcars[2:11]
P = 2 * ncol(X)
regType = 'lm'
family = 'gaussian'
seed = 1

initPop <- initialize(Y, X, P, regType, family, seed)
crossover(initPop[[1]]$variables, initPop[[2]]$variables, "single")
crossover(initPop[[1]]$variables, initPop[[2]]$variables, "multiple", 3)

}
