% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{support function: crossover()}
\usage{
crossover(ch1, ch2, nCuts = 1)
}
\arguments{
\item{ch1}{string; the first parent}

\item{ch2}{string; the second parent}

\item{nCuts}{integer; How many cuts do you want for the crossover? Defaults to 1}
}
\description{
This function takes for input 2 genes and creates 2 new genes from them.
}
\examples{
crossover()
}
\keyword{crossover}
