% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover}
\usage{
crossover(p1, p2, crossover_count = 1)
}
\arguments{
\item{p1}{vector; chromosome for Parent 1}

\item{p2}{vector; chromosome for Parent 2}

\item{crossover_count}{positive integer; Number of genes to be cross overed
(default set as 1)}
}
\value{
returns a list of two vectors corresponding to two offsprings
}
\description{
This function crosses over two chromosomes (referred as parent chormosomes) to
generate two offsprings. Default cross over count is 1.
}
