% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover}
\usage{
crossover(parents, parents_info)
}
\arguments{
\item{parents}{a n by 2 matrix consisting of the indices of the selected parents from the original dataframe}

\item{parents_info}{info matrix containing all the information of all parents}
}
\description{
This function takes genes from two parents based on the selected pool and produces a new offspring
}
\examples{
# Generate a random population and crossover to create a new population
population <- sample(100:200, 1)
n <- sample(10:20, 1)
chromosomes <- matrix(rbinom(population*n,1, 0.5), population, n)
pairs <- matrix(sample(1:population, 2*population, replace = TRUE), population, 2)
offsprings <- crossover(pairs, chromosomes)
}
\author{
Lulude Sun
}
