% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover Function for Genetic Algorithm for Variable Selection}
\usage{
crossover(P1, P2, C)
}
\arguments{
\item{P1}{First individual}

\item{P2}{Second individual}

\item{C}{The number of independent variables to be selected from}
}
\description{
Implements genetic algorithm crossover at randomly selected locus for two individuals in a generation
}
\examples{
# define initial parameters
C <- 5

# initiate parent generation
parents <- initiation( C = C , P = 30 )

# call crossover function
crossover( P1 = parents[ 1 , ] , P2 = parents[ 2 , ] , C = C )
}
\keyword{algorithm,}
\keyword{crossover}
\keyword{genetic}
\keyword{model}
\keyword{selection,}
