% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossOverMutate.R
\name{CrossOverMutate}
\alias{CrossOverMutate}
\title{Determine the fitness of some model}
\usage{
CrossOverMutate(generation, parent.index, prob.mutate)
}
\description{
This function takes a model object assesses its fitness
}
\examples{
CrossOverMutate()
}
