% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fitness.R
\name{evaluate_fitness}
\alias{evaluate_fitness}
\title{Function to evaluate fitness of chromosomes.}
\usage{
evaluate_fitness(generation_t0, Y, X, family, parallel, minimize,
  objective_function, rank_objective_function)
}
\arguments{
\item{generation_t0}{a matrix of parent chromosomes to be evaluated. Columns correspond to predictors/genes and rows correspond to parents/chromosomes.}

\item{Y}{vector of response variable}

\item{X}{a matrix or dataframe of predictor variables}

\item{family}{a character string describing the error distribution and link function to be used in the model. Default is gaussian.}

\item{rank_objective_function}{a function that ranks parents by their fitness as determined by optimize criteria.}

\item{calc_objective_function}{a function for optimizies \code{optimize}. Default is \code{\link{stats::AIC}}. User can specify custom function.}
}
\description{
This function uses default or user supplied ojective functions, rank functions, to evaluate fitness of a generation chromosomes. This process can be parallized if \code{parallel = TRUE}. Parallel operations uses \code{\link{mclapply}} to parallelize operations. Number of parallel operations is determined by \code{\link{detectCores} - 1}. Prescheduling = TRUE is only option for compuations. It is advised to only use parallelization if inputed a large dataframe with great than 1000 oberations and/or predictors.
}
