% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{support function: evaluation()}
\usage{
evaluation(X, Y, currentGeneration, popNum = 100, reg = "lm",
  criterion = "AIC")
}
\arguments{
\item{X}{is a matrix containing the variables in the model.}

\item{Y}{is a vector of observations to be used in the fitting process.}

\item{currentGeneration}{is the initialized generation and the selectively produced generations.}

\item{popNum}{What is the population size of chromosomes? Default to 100.}

\item{reg}{What is the regression type? Default to lm.}

\item{criterion}{What is the objective criterion function? Default to AIC.}
}
\description{
This function evaluate the fitness value of chromosomes.
}
\examples{
evaluation()
}
\keyword{evaluation}
