% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractBestIndividual.R
\name{ExtractBestIndividual}
\alias{ExtractBestIndividual}
\title{Determine the fitness of some model}
\usage{
ExtractBestIndividual(generation, fitnessmatrix)
}
\description{
Called from within GeneticAlgorithmFit
}
\examples{
ExtractBestIndividual()
}
