% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractResponseVariable.R
\name{ExtractResponseVariable}
\alias{ExtractResponseVariable}
\title{Extract response variable from dataframe}
\usage{
ExtractResponseVariable(dataset, name)
}
\arguments{
\item{dataset}{A matrix, datatable, or dataframe}
}
\description{
\code{ExtractResponseVariable()} allows the user to take a data table or dataframe
\code{dataset} and extract a 'response' vector indicated by \code{response.name}.  All
remaining columns are collated into a matrix of co-variates or 'predictors'.
}
\details{
\emph{ExtractResponseVariable()} returns the response vector and predictor matrix as a
  two-item list.

Called from within Select()
}
\examples{

\\code{\\link[GA]{Select}}

}
