% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.r
\docType{package}
\name{GA}
\alias{GA}
\alias{GA-package}
\title{R package: GA}
\description{
Implement a genetic algorithm for variable selection in regression problems.
}
\details{
User can provide matrix of predictors X and matrix of responses Y.
This GA package uses objective function, which defaults to AIC,
to find the fittest variable selection in regression problems.
}
