% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-new.R, R/selection.R
\name{gaLRselection}
\alias{gaLRselection}
\alias{gaLRselection}
\title{Parent Selection}
\usage{
gaLRselection(population, fitnessVec)

gaLRselection(population, fitnessVec)
}
\arguments{
\item{population}{a matrix}

\item{fitnessVec}{a vector of length n or a matrix with n rows}

\item{population}{a matrix}

\item{population}{A matrix}

\item{fitnessVec}{A vector}
}
\value{
returns the selected parents

The selected parents
}
\description{
Parent Selection

Parents Selection
}
\examples{

TBD
}
