% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaSelection.R
\name{gaSelection}
\alias{gaSelection}
\title{Parents Selection}
\usage{
gaSelection(methodFun, methodArgs)
}
\arguments{
\item{population}{a matrix}

\item{fitnessVec}{a vector of length n or a matrix with n rows}

\item{population}{a matrix}
}
\value{
returns the selected parents
}
\description{
This is a process for selecting parents which mate and recombine to create off-springs for the next generation.
}
\examples{

TBD
}
