% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R, R/gene_operator.R
\name{gene_selection}
\alias{gene_selection}
\alias{gene_selection}
\alias{gene_selection}
\title{Genetic operator: 1. Performs the k points crossover from 2 parents over
n_var genes at the indices given in the vector 'points'
2. Sample k locis and swap the alleles at these locis among the 2 parents.
3. Gene mutation at a specific rate}
\usage{
gene_selection(points, p1, p2, k, n_var, mu)

gene_selection(points, p1, p2, k, n_var, mu)

gene_selection(points, p1, p2, k, n_var, mu)
}
\arguments{
\item{points}{crossover points}

\item{p1, p2}{the 2 parents: lists with fields (variables, indices, linear_model)}

\item{k}{number of locis to be swapped}

\item{n_var}{length of the chromosome}

\item{points}{crossover points}

\item{p1, p2}{the 2 parents: lists with fields (variables, indices, linear_model)}

\item{k}{number of locis to be swapped}

\item{n_var}{length of the chromosome}

\item{points}{crossover points}

\item{p1, p2}{the 2 parents: lists with fields (variables, indices, linear_model)}

\item{k}{number of locis to be swapped}

\item{n_var}{length of the chromosome}
}
\description{
Genetic operator: 1. Performs the k points crossover from 2 parents over
n_var genes at the indices given in the vector 'points'
2. Sample k locis and swap the alleles at these locis among the 2 parents.
3. Gene mutation at a specific rate

Genetic operator: 1. Performs the k points crossover from 2 parents over
n_var genes at the indices given in the vector 'points'
2. Sample k locis and swap the alleles at these locis among the 2 parents.
3. Gene mutation at a specific rate

Genetic operator: 1. Performs the k points crossover from 2 parents over
n_var genes at the indices given in the vector 'points'
2. Sample k locis and swap the alleles at these locis among the 2 parents.
3. Gene mutation at a specific rate
}
