% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{generate_mutation}
\alias{generate_mutation}
\title{Generate offsprings after mutation}
\usage{
generate_mutation(input, mutation_rate, main_dataset)
}
\arguments{
\item{input}{A matrix contains all the populations which are required to mutate.
The number in each cell should be either 0 or 1. Each row represents an individual.
Each column represents the variable.}

\item{mutation_rate}{The chance for each variable in each individual to change the value.}

\item{main_dataset}{The dataset used to run the regression model.
And its model help to get the corresponding value of goodness of fit.}

\item{regression_target}{character: column of df regressed on other covariates in df}
}
\description{
generate_mutation is used to generate offsprings after mutations based on particular mutation rate.
And it will give the values of goodness of fit for each offspring.
}
\examples{
# Generate random input and choose a specific dataset in R
regression_target <- "V1"
toy_datasets <- generate_toy_dataset(regression_target = regression_target)
main_dataset <- toy_datasets$main_dataset
population <- as.matrix(toy_datasets$population)
generate_mutation(population, 0.01, main_dataset,regression_target)
}
\author{
Xiao Li
}
