% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R, R/util.R
\name{get_largest_interactions}
\alias{get_largest_interactions}
\alias{get_largest_interactions}
\alias{get_largest_interactions}
\title{Usded to get the c}
\usage{
get_largest_interactions(dataset, y, nb.var, max = 0)

get_largest_interactions(dataset, y, nb.var, max = 0)

get_largest_interactions(dataset, y, nb.var, max = 0)
}
\arguments{
\item{dataset}{the dataframe containing the data}

\item{y}{the response variable}

\item{nb.var}{the threshold p-value, a covariate is kept if its p-value is
less than p_val}

\item{dataset}{the dataframe containing the data}

\item{y}{the response variable}

\item{nb.var}{the threshold p-value, a covariate is kept if its p-value is
less than p_val}
}
\value{
a vector containing the names of the covariates that have been kept
################## PLEASE EDIT THE DOCUMENTATION PART ###################### 
Usded to get the c

a vector containing the names of the covariates that have been kept
################## PLEASE EDIT THE DOCUMENTATION PART ###################### 
Usded to get the c
}
\description{
Usded to get the c

Used to get the largest interaction terms.

Used to get the largest interaction terms.
}
