% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_prob_individuals}
\alias{get_prob_individuals}
\title{get_prob_individuals}
\usage{
get_prob_individuals(individuals, objective = "AIC")
}
\arguments{
\item{individuals}{list of (variables, indices, linear_model) see the
results section of random_selection_regression for more details}

\item{objective}{objective function}
}
\value{
a list of probabilities that correspond to the objective value.
}
\description{
Used to get the probabilities of each individual in a list
of (variables, indices, linear_model) in the next generation
proportionally to the value of their objective
}
