% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init}
\alias{init}
\title{utils.R}
\usage{
init(df, P, c)
}
\arguments{
\item{df}{(data.frame): the datasets with X and Y.}

\item{P}{(int): number of candidates per generations}

\item{c}{(int): number of genes per chromosome candidate}
}
\value{
generation(binary matrix P x c): P candidates
}
\description{
This is the R file containing all functions used for the genetic algorithm (GA)
package.  An OOP approach was taken to complete this algorithm.

function outputs P random binary vectors of length c
}
