% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a dataframe with P columns and C rows}
\usage{
init(P, C)
}
\arguments{
\item{P}{A number.}

\item{C}{A number.}
}
\value{
a dataframe with P columns and C rows.
}
\description{
Initialize a dataframe with P columns and C rows
}
\examples{
init(5, 6)
init(3, 4)
}
