% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main.R}
\usage{
main(P, c, collumn_to_use, k, eps, method, X, Y, fitness_function)
}
\arguments{
\item{P}{(int) number of candidates per generation}

\item{c}{(int) number of chromosomes per candidate solution}

\item{collumn_to_use}{Not sure what this is used for}

\item{k}{(int) total number of generations}

\item{eps}{(float) epsilon convergence condition, will exit algorithm if this condition is
satisfied before total number of generations is reached}

\item{X}{raw data to be fit}

\item{Y}{output values of function}

\item{fitness_function}{type of function used to compute the fitness of each candidate.
Either AIC or BIC is used in this package.}
}
\value{
model Final model output from running GA
}
\description{
This file calls the functions from utils to fit a model using genetic algorithm
Takes all of the components of the from utils and runs the full algorithm on the input data.
}
