% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutating Genes}
\usage{
mutate(childVector, mutationRate)
}
\arguments{
\item{childVector}{a vector representing the genes of child chromosomes}

\item{mutationRate}{mutation rate}
}
\description{
\code{mutate} takes in a vector of genes and try mutating each of them independently with a mutation rate
}
