% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{mutation function}
\usage{
mutation(chromosome, pMutation)
}
\arguments{
\item{chromosome}{bolean vector.}

\item{pMutation}{Mutation probability.}
}
\description{
This function allows you to mutate allels in a chromosome chain.
}
\examples{
mutation(chromosome,pMutation)
}
\keyword{cats}
