% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Selects regression variables via a genetic algorithm}
\usage{
mutation(population, mutation.prob = 0.1)
}
\arguments{
\item{population}{matrix of dimension n * p}

\item{mutation, prob}{mutation probability}
}
\value{
returns a matrix of dimension n * p
}
\description{
Selects regression variables via a genetic algorithm
}
\examples{

TBD
}
