% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{support function: mutation()}
\usage{
mutation(chromosome, probability = 0.01)
}
\arguments{
\item{chromosome}{vector; one chromosome with genes 0/1}

\item{probability}{numeric; below or equal to 1 and above or equal to 0; What is the probability of mutation? Defaults to 0.01}
}
\description{
This function mutate the chromosomes.
}
\examples{
mutation()
}
\keyword{mutation}
