% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{support function: mutation()}
\usage{
mutation(chromosome, probability = 0.01)
}
\arguments{
\item{chromosome}{is your initial set of chromosomes}

\item{probability}{What is the probability of mutation? Defaults to 0.01.}
}
\description{
This function mutate the chromosomes.
}
\examples{
mutation()
}
\keyword{mutation}
