% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation}
\usage{
mutation(chromosome, mut_prob = 0.05, mutations = 1)
}
\arguments{
\item{chromosome}{vector; an individual chromosome for mutation}

\item{mut_prob}{float; probability of mutation (between 0 and 1) (default is
.05)}

\item{mutations}{positive integer; number of genes to be mutated if selected
to mutate (default set to 1)}
}
\value{
returns a vector after mutation
}
\description{
The mutation function takes the vector of a chromsome and introduce random
mutations into it. Mutations occur with a probability determined by the user
and if an offspring is selected for mutation, user determines the number
of genes that will be mutated.
}
