% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation
Calculate the fitness of each individual in the population}
\usage{
mutation(mat, mu)
}
\arguments{
\item{mat}{an n by p matrix of offsprings, where n is the number of individuals and p is the length of the gene}

\item{mu}{numeric, the probability of mutation}
}
\value{
returns the mutated offspring population matrix
}
\description{
Mutation
Calculate the fitness of each individual in the population
}
\examples{
# create matrix, and perform mutation
x <- matrix(rep(c(1,0),5),5,2)
# perform mutation
mutation(mat=x,mu=0.5)
}
\author{
Wei-Hsiang Sun
}
