% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation}
\usage{
mutation(parent_1, parent_2, offspring, m)
}
\arguments{
\item{parent_1, parent_2}{Numeric vectors with binary 0/1 elements. Should be of same length.}

\item{offspring}{A Numeric vector generated by \code{crossover} and/or other genetic operators.}

\item{m}{Mutation rate: a numeric value between 0 and 1. By default 0.01.}
}
\description{
\code{mutation} is to do mutation for the offspring after breeding.
}
\details{
Changes an offspring chromosome by randomly introducing one or
more alleles in loci where those alleles are not seen in the
corresponding loci of either parent chromosome.
}
