% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popInitial.R
\name{popInitial}
\alias{popInitial}
\title{support function: popInitial()}
\usage{
popInitial(popNum = 100, geneLength, zeroRate = 0.5)
}
\arguments{
\item{popNum}{integer; What is the population size of chromosomes? Default to 100}

\item{geneLength}{integer; number of genes (0/1) on a chromosome}

\item{zeroRate}{numeric; below or equal to 1 and above or equal to 0; What is the rate of zero among the genes of the first generation? Defaults to 0.5}
}
\description{
This function intializes the population.
}
\examples{
popInitial()
}
\keyword{popInitial}
