% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSelection.R
\name{rankSelection}
\alias{rankSelection}
\title{support function: rankSelection()}
\usage{
rankSelection(dfCurrent, usingRank = TRUE)
}
\arguments{
\item{dfCurrent}{dataframe; one column corresponding to the genes (string), and one to the AIC values of each chromosome}

\item{usingRank}{logic; TRUE define selection probabilities using the rank and FALSE create selection probabilities proportionate to fitness; defaults to TRUE}
}
\description{
This function create selection probabilities of the chromosomes.
}
\examples{
rankSelection()
}
\keyword{rankSelection}
