% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{rankselection}
\alias{rankselection}
\title{Rank based selection}
\usage{
rankselection(fitness_rank)
}
\arguments{
\item{fitness_rank}{fitness rank for individual solutions}
}
\value{
Returns a matrix pop_size by 2 documenting the row index of each parent in population
}
\description{
Select one parent with probability proportional to reverse fitness_rank
Select the other parent randomly
Iterate this process for pop_size times
}
\examples{
rankselection(fitness_rank = c(1, 6, 5, 9, 10, 7, 2, 3, 8, 4))
}
\author{
Chloe Zhang
}
