% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Runs regression using a subset of variables dictated by the genotype}
\usage{
regress(genotype, x, y, model = "glm", fitnessCriteria = "AIC",
  modelParams = NULL)
}
\arguments{
\item{genotype}{array of length p made up of 0s and 1s; represents a single member of the population}

\item{x}{matrix of dimension n * p}

\item{y}{vector of length n or a matrix with n rows}

\item{model}{vector specifying the model: the first argument should be "lm" or "glm", and subsequent arguments specify additional arguments into "lm" or "glm"}

\item{fitnessCriteria}{default "AIC", a string specifying the fitness criterion: "AIC", "BIC", or an attribute of a fitted lm or glm model (must be single numeric value)}
}
\description{
Runs regression using a subset of variables dictated by the genotype
}
