% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/Utilities.R, R/regression.R
\name{regression}
\alias{regression}
\alias{regression}
\alias{regression}
\title{Performs a linear regression given y, a list of covariates and the
indices of the covariates to take into account.}
\usage{
regression(y, names, indices, dataset, reg_method)

regression(y, names, indices, dataset, reg_method)

regression(y, names, indices, dataset, reg_method)
}
\arguments{
\item{y}{Name of the response variable in the dataset}

\item{names}{Names of the covariates in the dataset}

\item{indices}{Indices of the covariates considered in the regression}

\item{dataset}{Name of the dataset}

\item{reg_method}{Regression method}

\item{y}{Name of the response variable in the dataset}

\item{names}{Names of the covariates in the dataset}

\item{indices}{Indices of the covariates considered in the regression}

\item{dataset}{Name of the dataset}

\item{y}{Name of the response variable in the dataset}

\item{names}{Names of the covariates in the dataset}

\item{indices}{Indices of the covariates considered in the regression}

\item{dataset}{Name of the dataset}

\item{reg_method}{Regression method}
}
\value{
'lm' object obtained by performing the regression with the
 selected covariates

'lm' object obtained by performing the regression with the
 selected covariates

'lm' object obtained by performing the regression with the
 selected covariates
}
\description{
Performs a linear regression given y, a list of covariates and the
indices of the covariates to take into account.

This is the R file containing all functions used for the genetic algorithm (GA)
package.  An OOP approach was taken to complete this algorithm.
Performs a linear regression given y, a list of covariates and the
indices of the covariates to take into account.

Performs an lm or glm given y, a list of covariates and the
indices of the covariates to take into account.
}
