% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceClones.R
\name{ReplaceClones}
\alias{ReplaceClones}
\title{Determine the fitness of some model}
\usage{
ReplaceClones(generation, fitness.vec, C)
}
\description{
Called from within GeneticAlgorithmFit
}
\examples{
ReplaceClones()
}
