% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{{select()}}
\usage{
select(df, P, max_iter, method_text, fitness_function_text, mu,
crossover_points, mechanism, random = TRUE, Gap = 1/4, plot.return = FALSE)
}
\arguments{
\item{df}{(data frame) Dataset to fit}

\item{P}{(int) The number of individuals per generation.}

\item{max_iter}{(int) The maximum number of iterations allowed when running GA}

\item{method_text}{lm() or glm() methods for fitting the data}

\item{fitness_function_text}{AIC, BIC, ow}

\item{random}{(logical) Random replacement on or off for parent selection}

\item{Gap}{Generation gap that determines how parents (generation t) are replaced by offspring
of the (t+1) generation}

\item{plot.return}{(logical) Boolean for returning plot at end of the algorithm}

\item{mu:}{Mutation rate of the for each allele within the a given candidate chromosome.}

\item{crossover_points:}{(int) The number of crossover points during breeding step}

\item{mechanism:}{The mechanism to selection parents by.  Selection mechanisms
are "ranked" or "tournament."}
}
\description{
This is main call function to run package GA.  This package is comprised of
a main execution file (\emph{select.R} and a R file comtaining all functions
that are necessary for execution (\emph{utils.R}).  The user can enter in a dataset and provide
variables (listed below) to execute the genetic algorithm.
}
\details{
Contained in the list below are the invdividual functions that are called during the
execution of the genetic algorithm.
\itemize{
 \item{init()}: {Initializes the dataframe for the dataset}
 \item{training()}: {fits the method on candidates and return the fitness value of the candidate}
 \item{select_parents()}: {Chooses parents for breeding based off of ranked
 or tournament selection using the fitness values of each parent}
 \item{breed()}: {Breeds parents based off of pairing from select_parents()}
 \item{crossover()}: {Function within breed() to conduct crossover between parent pairs
 during breeding step}
 \item{mutation()}: {Determines if alleles of offspring (t+1 generation) mutates or not}
 \item{get_model()}: {Returns the best fit model of the dataset}
}
}
