% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selects Regression Variables using a Genetic Algorithm}
\usage{
select(x, y, model = list("lm"), fitMetric = "AIC", maxGen = 200L,
  minGen = 10L, gaMethod = list("TN", 5), pop = 100L, pMutate = 0.1,
  crossParams = c(0.8, 1L), eliteRate = 0.1, ...)
}
\arguments{
\item{x}{matrix of dimension n * p}

\item{y}{vector of length n or a matrix with n rows}

\item{model}{list - default "lm" : one of ("lm", "glm") and an optional character string specifying arguments into lm.fit() or glm.fit()}

\item{fitMetric}{default "AIC": one of ("AIC", "BIC", "RSS") or a function that takes a regression object and outputs a single number to be maximized}

\item{maxGen}{default 200: integer specifying the maximum number of GA generations to use}

\item{minGen}{default 10: integer specifying the number of generations without fitness improvement at which the GA algorithm will stop}

\item{gaMethod}{list - default 'LR': one of ('TN', 'LR', 'ER','RW') and an additional numrical argument as needed. See gaSelection for details.}

\item{pop}{default 100: integer specifying the size of the genotype pool.}

\item{pMutate}{default .1: real number between 0 and 1 specifying the probability of an allele mutation}

\item{crossParams}{numeric - default (.8, 1): c("cross probability", "max number of cross locations on a single gene")}

\item{eliteRate}{default .1: Proportion of highest fitness genotypes that pass into the next generation unchanged.}
}
\value{
returns a list of 4 components: optimum, fitPlot, fitStats, and GA
\itemize{
\item{\strong{optimum:} a list of properties of the genotype acheiving max fitness}{
\itemize{
\item \emph{\strong{variables:}} the recommended set of regression variables
\item \emph{\strong{fitness:}} the achieved fitness metric
\item \emph{\strong{fitModel:}} the regression object returned by using the recommended variables
}}
\item{\strong{fitPlot:} a plot of the mean, median, and maximum fitness over the generations}
\item{\strong{fitStats:} a tibble of the values used to generate the plot}
\item{\strong{GA:} a list of data associated with each generation of the genetic algorithm}{
\itemize{
\item \emph{\strong{fitness:}} the fitness measures of the current generation
\item \emph{\strong{elites:}} the fitness values and genotypes with the highest fitness
}}
}
}
\description{
Selects Regression Variables using a Genetic Algorithm
}
\examples{
tbd
}
\seealso{
\code{\link{regress}}
\code{\link{gaSelection}}
\code{\link{evolve}}
}
