% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selects regression variables via a genetic algorithm}
\usage{
select(obj)
}
\arguments{
\item{x}{matrix of dimension n * p}

\item{y}{vector of length n or a matrix with n rows}

\item{model}{a list: first entry specifying either"lm" or "glm", and subsequent named entries specifying additional arguments into lm.fit() or glm.fit(): default "glm"}

\item{fitMetric}{"AIC", "BIC", or "RSS": default "AIC"}

\item{maxGen}{an integer specifying the maximum number of GA generations to use: default 100}

\item{minGen}{an integer specifying the number of generations without fitness improvement at which the GA algorithm will stop: default 5}

\item{gaMethod}{a list: The first entry must be one of ('TN', 'LR', 'ER','RW') with other entries specifying required additional arguments as needed. See gaMethod help for details.}

\item{pop}{an integer specifying the size of the genotype population. If odd, 1 random individual will be left out of crossing: default 100}

\item{pMutate}{a real value between 0 and 1 specifying the probability of an allele mutation: default .1}

\item{crossParams}{a numeric vector, c("cross probability", "max number of cross locations on a single gene"): default (.8, 1)}

\item{eliteRate}{Proportion of total population of elites who pass into the next generation unchanged}

\item{a}{A placeholder.}

\item{a}{A placeholder.}

\item{a}{A placeholder.}

\item{a}{A placeholder.}

\item{a}{A placeholder.}

\item{...}{optional arguments to lm, glm, ect.}

\item{coefficients}{a named vector of coefficients}

\item{fitness}{the maximum value attained of the specified fitness criterion}

\item{generations}{the number of GA generations}
}
\value{
returns an object of class "GA", which is a list containing the following components:
}
\description{
Selects regression variables via a genetic algorithm
}
\examples{

TBD
}
