% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{select}
\usage{
select(dat, y_name, popSize = 30, pSelect = 0.2, pMutate = 0.01,
  max_iter = 1000, fitfunc = "AIC", family = "gaussian",
  end_iter_cond = F)
}
\arguments{
\item{dat}{dataframe, data to operation(default sets dependent variable in first column and independent varialbes is other columns)}

\item{y_name}{string, column of dependent variable}

\item{popSize}{integer, population size(default=30)}

\item{pSelect}{float, the worse part to remove from the population(default=0.2)}

\item{pMutate}{float, the probability used to decide number of mutation(default=0.01)}

\item{max_iter}{integer, the maximum number of iteration}

\item{fitfunc}{method, model selection method(default is AIC)}

\item{family}{family,for linear regression model the family should be a continuous probability density function (default is gaussian family)}

\item{end_iter_cond}{logical, the index to decide whether to stop iteration (default=F)}
}
\value{
list, return the best model that composed of chromsome and corresponding fitness
}
\description{
Select the best model
}
