% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Identify the most important variables for prediction}
\usage{
select(X, y, num_iter = 100)
}
\arguments{
\item{y}{A response vector.}

\item{x}{An input matrix.}
}
\value{
A list of the features of \code{X} that best describe \code{y}.
}
\description{
Identify the most important variables for prediction
}
\examples{
add(1, 1)
add(10, 1)
}
