% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{support function: select()}
\usage{
select(X, Y, popNum = 100, reg = "lm", criterion = "AIC",
  useParallel = FALSE, numCores = 4, usingRank = TRUE, cross_cutNum = 1,
  mutation_prob = 0.01, mut_pCurve = FALSE, initial_zeroRate = 0.5,
  min_iter = 10, max_iter = 500)
}
\arguments{
\item{X}{matrix/dataframe; predictors}

\item{Y}{vector/matrix/dataframe; observations to be used in the fitting process}

\item{popNum}{integer; What is the population size of chromosomes? Default to 100}

\item{reg}{string; 'lm'/'glm; What is the regression type? Default to 'lm'}

\item{criterion}{string; function name; What is the objective criterion function? Default to 'AIC'}

\item{useParallel}{logic; Should it use parallelization? Default to FALSE}

\item{numCores}{integer; How many cores should it use? Default to 4}

\item{usingRank}{logic; TRUE define selection probabilities using the rank and FALSE create selection probabilities proportionate to fitness; defaults to TRUE}

\item{cross_cutNum}{integer; How many cuts do you want for the crossover? Defaults to 1}

\item{mutation_prob}{numeric; below or equal to 1 and above or equal to 0; What is the probability of mutation? Defaults to 0.01}

\item{mut_pCurve}{logic; Should it use curve for the probability of mutation (Starting with a big probability and decreasing to mutation_prob by the time)? Defaults to FALSE.}

\item{initial_zeroRate}{numeric; below or equal to 1 and above or equal to 0; What is the rate of zero among the genes of the first generation? Defaults to 0.5}

\item{min_iter}{integer; What is the minimum number of iterations? Defaults to 10.}

\item{max_iter}{integer; What is the maximum number of iterations? Defaults to 500.}
}
\description{
This is the primary function of the R package of this project.
}
\examples{
select()
}
\keyword{select}
