% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{support function: select()}
\usage{
select(X, Y, popNum = 100, reg = "lm", criterion = "AIC",
  useParallel = FALSE, numCores = 4, usingRank = TRUE, cross_cutNum = 1,
  mutation_prob = 0.01, mut_pCurve = FALSE, initial_zeroRate = 0.5,
  min_iter = 10, max_iter = 500)
}
\arguments{
\item{X}{is a matrix containing the variables in the model.}

\item{Y}{is a vector of observations to be used in the fitting process.}

\item{popNum}{What is the population size of chromosomes? Default to 100.}

\item{reg}{What is the regression type? Default to lm.}

\item{criterion}{What is the objective criterion function? Default to AIC.}

\item{useParallel}{Should it use parallelization? Default to FALSE.}

\item{numCores}{How many cores should it use? Default to 4.}

\item{usingRank}{TRUE define selection probabilities using the rank and FALSE create selection probabilities proportionate to fitness. Defaults to TRUE.}

\item{cross_cutNum}{How many cuts do you want for the crossover? Defaults to ONE.}

\item{mutation_prob}{What is the probability of mutation? Defaults to 0.01.}

\item{mut_pCurve}{Should it use curve for the probability of mutation (Starting with a big probability and decreasing to mutation_prob by the time)? Defaults to FALSE.}

\item{initial_zeroRate}{What is the rate of zero among the genes of the first generation? Defaults to 0.5.}

\item{min_iter}{What is the minimum number of iterations? Defaults to 10.}

\item{max_iter}{What is the maximum number of iterations? Defaults to 500.}
}
\description{
This is the primary function of the R package of this project.
}
\examples{
select()
}
\keyword{select}
