% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Primary Select Function}
\usage{
select(
  Y,
  X,
  models,
  core = 1,
  criteria = "AIC",
  f = 0.5,
  fselect = "standard",
  crossover_count = 1,
  mut_prob = 0.01,
  mutations = 5,
  pop_size = 200,
  converge = "delta",
  num_iter = 100,
  reg_model = lm
)
}
\arguments{
\item{Y}{vector; target variable}

\item{X}{matrix/dataframe; design matrix (predictors)}

\item{models}{matrix; current generation of individuals (candidate models)}

\item{core}{positive integer; No of cores to be used for parallization
(default 1 which means no parallization)}

\item{criteria}{string; objective criteria to be used}

\item{f}{float; fraction of indiviuals to be selected from the current
generation during fitness selection}

\item{fselect}{string; selection criteria to be used}

\item{crossover_count}{positive integer; Number of genes to be cross overed
(default set as 1)}

\item{mut_prob}{float; probability of mutation (between 0 and 1) (default is
.05))}

\item{mutations}{positive integer; number of genes to be mutated if selected
to mutate (default set to 1)}

\item{pop_size}{positive integer; size of the original population (no of
individuals)}

\item{converge}{string; converging criteria to be used (should be count or
delta)}

\item{num_iter}{positive integer; maximum number of iterations}

\item{reg_model}{model type (lm or glm) default is lm}
}
\value{
returns a list of variables best fitted for the problem
}
\description{
Main function that uses all the other auxillary functions to select the
best variables for a given regression problem.
}
\examples{

library(tidyverse)
library(GA)

#let's simulate some data to test our algorithm on

set.seed(121119)
x <- matrix(rnorm(2000*100, 0,5), nrow=2000, ncol=100)

#Let's simulate 5 coefficients to be non-zero
beta0 <- 2
beta1 <- 3
beta10 <- 1
beta60 <- 6
beta92 <- 9
beta87 <- 2

#Simulate sample y (output variable)
y <- beta0 + beta1*x[,1] + beta10 * x[,10] + beta60*x[,60] + beta92*x[,92] + beta87*x[,87] + rnorm(2000)

#Calling the Select function with default arguments
select(y,x)

#Calling the select function with user defined inputs
library(doParallel)
library(future)
select(y,x,core=4,criteria="BIC",f=0.6,crossover_count=3,mut_prob=0.05, num_iter=500,reg_model=glm)

}
