% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selects covariates for regression using a genetic algorithm}
\usage{
select(dataset, regression, criterion = "AIC", parent_selection = "fitness",
  gene_selection = "default", num_iter = 20)
}
\arguments{
\item{dataset}{The dataset in matrix form with last column being the dependent variable.}

\item{regression}{The regression form to use (either lm or glm).}

\item{criterion}{The objective criterion to use (default AIC).}

\item{parent_selection}{The selection method for choosing parents in GA.}

\item{gene_selection}{The selection method for choosing genes in GA.}

\item{num_iter}{The maximum number of iterations for the algorithm}
}
\value{
List containing the fitted regression, the optimal objective value,
and number of iterations used.
}
\description{
Selects covariates for regression using a genetic algorithm
}
\examples{

}
