% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{select}
\alias{select}
\title{Genetic Algorithm for Model Selection}
\usage{
select(y, dataset, reg_method = NULL, n_iter = 200, pop_size = 2 * n, objective = "AIC",
interaction = F, most_sig = F, parent_selection = "prop", nb_groups = 4, generation_gap = 0.25,
gene_selection = "crossover", nb_pts = 1, mu = 0.3, err = 1e-6)
}
\arguments{
\item{y}{(character) Column name of the dependent variable}

\item{dataset}{(data frame)The dataset in matrix form with last column being the dependent variable.}

\item{reg_method}{(character) "lm" or "glm". methods for fitting the data}

\item{n_iter}{(int) The maximum number of iterations allowed when running GA}

\item{pop_size}{(int) The number of individuals per generation.}

\item{objective}{(character) The objective criterion to use (default AIC).}

\item{interaction}{(logical) Whether to add the interaction terms to the independent variables.}

\item{most_sig}{(logical) Whether to use the most significant variables inside the first_generation function.}

\item{parent_selection}{(character) The mechanism to select parents. Selection mechanisms are "prop","prop_random", "random" or "tournament".}

\item{nb_groups}{(int) The number of groups chosen to do using the tournament selection.}

\item{generation_gap}{( numeric) The proportion of the individuals to be replaced by offspring.}

\item{gene_selection}{(function) The additional selection method for choosing genes in GA.
refer to gene_selection to see the required inputs and the desired form of output}

\item{gene_operator}{If the user doesn't provide his own gene_selection method. Then the gene_operator is used.}

\item{nb_pts}{(int) The number of points that used in crossover}

\item{mu}{(numeric) The mutation rate}

\item{err}{(numeric) The convergence threshold (if the difference between last iteration and current is
less than err, the algorithm stops)}
}
\value{
\code{select} returns a list with elements:
\itemize{
 \item{\code{count}}: {number of iterations until getting the selection}
 \item{\code{variables}}: {The names of variables that selected}
 \item{\code{model}}: {a \code{lm} or \code{glm} object}
 \item{\code{fitness_value}}: {the value of fitness function of the returned model}
 }
}
\description{
This is main call function to run package GA.  This package is comprised of
a main execution file (\code{select.R}) and a R file comtaining all utilities functions
for execution (\code{utilities.R}).  The user can enter enter a dependent variable and  a
dataset to execute this function.
}
\examples{
select("mpg", mtcars)
select("crim", Boston)

}
