% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/MainFunction.R, R/select.R
\name{select}
\alias{select}
\alias{select}
\alias{select}
\title{Selects covariates for regression using a genetic algorithm}
\usage{
select(y, dataset, n_iter = 200, pop_size = 20, objective = "AIC",
  interaction = F, most_sig = F, normalization = F, selection = "prop",
  nb_groups = 4, permutation = 1, gene_selection = "crossover",
  nb_pts = 1, decay = "unif")

select(y, dataset, reg_method = NULL, n_iter = 200, pop_size = 2 * n, objective = "AIC",
interaction = F, most_sig = F, parent_selection = "prop", nb_groups = 4, generation_gap = 0.25,
gene_selection = "crossover", nb_pts = 1, mu = 0.3, err = 1e-6)

select(y, dataset, n_iter = 200, pop_size = 20, objective = "AIC",
  interaction = F, most_sig = F, normalization = F, selection = "prop",
  nb_groups = 4, permutation = 1, gene_selection = "crossover",
  nb_pts = 1, decay = "unif")
}
\arguments{
\item{y}{(character) Column name of the dependent variable}

\item{dataset}{The dataset in matrix form with last column being the dependent variable.}

\item{n_iter}{(int) The maximum number of iterations allowed when running GA}

\item{pop_size}{(int) The number of individuals per generation.}

\item{objective}{(character) The objective criterion to use (default AIC).}

\item{interaction}{(logical) Whether to add the interaction terms to the independent variables.}

\item{most_sig}{(logical) Whether to use the most significant variables inside the first_generation function.}

\item{nb_groups}{(int) The number of groups chosen to do using the tournament selection.}

\item{gene_selection}{The selection method for choosing genes in GA.}

\item{nb_pts}{(int) The number of points that used in crossover}

\item{regression}{The regression form to use (either lm or glm).}

\item{criterion}{The objective criterion to use (default AIC).}

\item{parent_selection}{The selection method for choosing parents in GA.}

\item{num_iter}{The maximum number of iterations for the algorithm}

\item{generation_gap}{The number of the individuals to be replaced}

\item{mu}{The mutation rate}

\item{dataset}{(data frame)The dataset in matrix form with last column being the dependent variable.}

\item{reg_method}{(character) "lm" or "glm". methods for fitting the data}

\item{parent_selection}{(character) The mechanism to select parents. Selection mechanisms are "prop","prop_random" or "tournament".}

\item{generation_gap}{( numeric) The proportion of the individuals to be replaced by offspring.}

\item{gene_selection}{(function) The additional selection method for choosing genes in GA.}

\item{mu}{(numeric) The mutation rate}

\item{err}{(numeric) The convergence threshold (if the difference between last iteration and current is
less than err, the algorithm stops)}

\item{dataset}{The dataset in matrix form with last column being the dependent variable.}

\item{regression}{The regression form to use (either lm or glm).}

\item{criterion}{The objective criterion to use (default AIC).}

\item{parent_selection}{The selection method for choosing parents in GA.}

\item{gene_selection}{The selection method for choosing genes in GA.}

\item{num_iter}{The maximum number of iterations for the algorithm}
}
\value{
List containing the fitted regression, the optimal objective value,
and number of iterations used.

\code{select} returns a list with elements:
\itemize{
 \item{\code{count}}: {number of iterations until getting the selection}
 \item{\code{variables}}: {The names of variables that selected}
 \item{\code{model}}: {a \code{lm} or \code{glm} object}
 \item{\code{fitness_value}}: {the value of fitness function of the returned model}
 }

List containing the fitted regression, the optimal objective value,
and number of iterations used.
}
\description{
Selects covariates for regression using a genetic algorithm

This is main call function to run package GA.  This package is comprised of
a main execution file (\code{select.R}) and a R file comtaining all utilities functions
for execution (\code{utilities.R}).  The user can enter enter a dependent variable and  a
dataset to execute this function.

Selects covariates for regression using a genetic algorithm
}
\examples{

select("mpg", mtcars)
select("crim", Boston)


}
