% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Genetic Algorithm based Variable Selection}
\usage{
select(y, x, family, k, P, mutation = 0.01, ncores = 0,
  fitness_function = stats::AIC, fitness = "rank", selection = "fitness",
  tol = 5e-04, maxIter = 100L)
}
\arguments{
\item{y}{a vector or matrix of responses}

\item{x}{a matrix of covariates}

\item{family}{a description of the error distribution to be used in the glm fitting. Default is to use gaussian.}

\item{P}{Population size, corresponding to the number of genes or covariates on each chromosome or model.}

\item{mutation}{an optional value specifying the rate at which mutation occurs in the new population.}

\item{ncores}{an optional value indicating the number of cores to use in parallelization. Should be numeric.}

\item{fitness_function}{optional function to evaluate model fitness. Must be of type closure. Default is to use AIC. Options: "rank" uses relative rank of models based on AIC. Option: "weight" uses absolute value of AIC to determine probability of reproduction in the preceding generation. This option should be used with caution because it can become stuck at a local minimum, as a single model with very low AIC will have large probability of reproduction.}

\item{tol}{Optional value indicating relative convergence tolerance. Should be of class numeric.}

\item{maxIter}{Optional value indicating the maximum number of iterations. Default is 100.
\itemize{
\item References
}
\enumerate{
\item Burnham, K. P. and D. R. Anderson. 2002. Model Selection and Multimodel Inference. Springer-Verlag, New York
\item Geof H. Givens, Jennifer A. Hoeting (2013) Combinatorial Optimization (italicize). Chapter 3 of Computational Statistics (italicize).
}}
}
\description{
Select utilizes the genetic algoritim to optimize model selection for generalized linear models.
Relying on functions: \code{\link{selection}}, \code{\link{crossover}} and \code{\link{mutation}}, new populations of chromosomes correspondng to models are generated, selecting based on AIC, until the optimal model is achieved.
}
\examples{

}
