% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{Select}
\alias{Select}
\title{Apply a genetic algorithm to find the optimal co-variates in a linear regression}
\usage{
Select(dataset, response.name, userfunc = "AIC", user.family = "gaussian",
  flag.log.scale = TRUE, frac.replace = 0.2, Niter = 100, Nruns = 1,
  mutate.rate = FALSE, plot.flag = TRUE)
}
\arguments{
\item{dataset}{A matrix, datatable, or dataframe}

\item{response.name}{The name of the column in \strong{dataset} that will act as the
response variable to be predicted}

\item{userfunc}{A fitness function that operates on a model that could be provided by the user.
The default is the Aikake Information Criteria or "AIC".}

\item{user.family}{Model family name to be passed to \code{glm}. Default is "gaussian"}

\item{flag.log.scale}{Default is TRUE if the log of the predictor varaible is to be fit.}

\item{frac.replace}{Fraction of worst parents to be replaced with the best children
in each generation}

\item{Niter}{Maximum number of iterations during each run. Default is 100}

\item{Nruns}{Number of times genetic algorithm is run.  Default is 1}

\item{mutate.rate}{Genetic algorithm mutation rate. If set to FALSE it is automatically
determined. A value of 0.01 is suggested}

\item{plot.flag}{Set to TRUE to plot the evolution of the population of individuals over the
  progression of the algorithm

\value{ \code{Select()} produces a single list each of whose elements are a list.vector of objects
  partaining  to each of the \code{Nruns} times that the genetic algorithm was run.  Each
  list element is named according the type of data it contains:
  \itemize{
    \item \code{LastGen} is a list of the last generation of solutions produced during each run
    \item \code{Fitness} is a list of the complete fitness matrix representing all generations
      during each run
    \item \code{BestModel} is a list of the best model produced during each run
    \item \code{BestFitness} is a vector of the fitness of the best model from each run
  }
}}
}
\description{
\code{Select()} is used to apply the genetic algorithm to some input dataset
outputs a list containing a list of the indiivdials in the final generation'
of the genetic algorithm, a matrix of the fitness values of all individuals in all
generations and the fitted model from glm for the best individual
}
\examples{
baseball = read.table(file.choose(),header=TRUE)
out <- Select(dataset=baseball, response.name="salary", Niter=50, Nruns=1, mutate.rate = 0.01)

}
\keyword{algorithm}
\keyword{genetic}
\keyword{linear}
\keyword{optimization}
\keyword{regression}
