% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select
A function that outputs the best subset of variables according to the aic criteria. User can select different fitness functions as well as different types of selection techniques.}
\source{
(".R/Mutation.R")

(".R/Fitness.R")

(".R/selection.R")

(".R/crossover.R")

(".R/Initialization.R")
}
\usage{
select(
  datax,
  datay,
  p,
  mu = 0.01,
  selection_type = "twoprop",
  model_type = "lm",
  max_iter = 100,
  fitness_func = fitness_aic,
  k = NULL,
  threshold = 10^-5
)
}
\arguments{
\item{datax}{a n by m matrix, where n is the number of observations, and m is the total number of variables in the data}

\item{datay}{a length n numeric vector where n is the number of observations in the data}

\item{p}{numeric, the size of population}

\item{mu}{numeric, has value between 0 and 1, mutation probability, is set to 0.01 by default}

\item{selection_type}{character, the type of selection to use, 'twoprop' for proportional selection for both parents, 'oneprop' for proportional selection for one parent, 'tournament' for tournament selection, 'rank' for rank based selection, 'stochastic' for random selection. by default, the selection_type is set to 'twoprop'}

\item{max_iter}{integer, maximum iteration, is set to 100 by default}

\item{k}{integer, groups for tournament selection}

\item{threshold}{numeric, the converging threshold, the criteria for convergence between and within population, threshold is set to 10^-5 by default}

\item{type}{character, the type of model to use, 'lm' for linear regression, 'glm' for logistic regression}
}
\value{
A list of two elements, the best aic value, and the best subset of variables. Variables selected corresponds to 1 in a numeric vector, while unselected variables are indexed 0.
}
\description{
Select
A function that outputs the best subset of variables according to the aic criteria. User can select different fitness functions as well as different types of selection techniques.
}
\examples{
# import library to create toy data for glm example
library("VSURF")
data("toys")
set.seed(1)
x <- toys[['x']]
y <- toys[['y']]

# selection results
select(x,y,p=10,selection_type = 'tournament',max_iter = 100,model_type = 'glm',k=5,threshold = 10^-3 )

# create data for lm example
set.seed(5)
x <- rnorm(1000,3,2)
dim(x) <- c(100,10)
y <- 3*x[,1] + 4*x[,7]- 5*x[,9] + rnorm(100)

# selection results
select(x,y,p=10,selection_type = 'tournament',max_iter = 100,k=5,threshold = 10^-4 )

}
\author{
Wei-Hsiang Sun
}
