% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Genetic Algorithm for Feature Selection}
\usage{
select(data, yColumn, fitnessFUN = match.fun(AIC), k = 2,
  geneticOperator = match.fun(singlePtCrossover),
  family = match.fun(gaussian), nIterations = 200)
}
\arguments{
\item{data}{dataframe containing dependent variable (y) and all covariates}

\item{yColumn}{integer specifying which column of data is dependent variable}

\item{fitnessFUN}{fitness function to use. Default is AIC}

\item{k}{selective pressure for ranking. Default is 2}

\item{geneticOperator}{genetic operator to use. Default is single point crossover.
The function must take 2 parents and the number of features as
input and return two children}

\item{family}{a description of the error distribution and
link function to be used in the model. Default is gaussian (standard lm()). To use
glm(), specify a different family such as poisson(), etc...}

\item{nIterations}{integer specifying number of iterations before stopping}
}
\value{
Binary vector detailing optimal features to use
}
\description{
Genetic Algorithm for Feature Selection. User must pass in a dataframe and integer
specifying which column to use as the dependent variable. The function also allows
for different fitness functions, genetic operators, and selective pressure to be applied
To use GLM, specific a non-gaussian family.
}
\examples{
bestFeatures <- select(data = fakeData, yColumn = 1, fitnessFUN = AIC, k = 2,
geneticOperator = singlePtCrossover, family = gaussian)
}
