% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{select Function}
\usage{
select(X, Y, ObjectiveFunction = AIC, Probs = Ranking, P = 2 * ncol(X),
  Initialize = Initialize(ncol(X), P), mu = 1/ncol(X), Stop = Stop,
  Iterations, ...)
}
\arguments{
\item{X}{A matrix of numbers.}

\item{Y}{A vector or matrix of numbers.}

\item{ObjectiveFunction}{A function. Defaults to AIC.}

\item{Probs}{A function. Defaults to Ranking.}

\item{P}{An integer. Must be even.}

\item{Initialize}{A vector or matrix.}

\item{mu}{The mutation rate has to be a number between 0 and 1.}

\item{Stop}{A function.}

\item{Iterations}{An integer.}
}
\value{
Return the fittest individual in the population.
}
\description{
This function implement a genetic algorithm for variable selection in regression problems.
}
\examples{
select()

}
