% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{select Function}
\usage{
select(X, Y, ObjectiveFunction = AIC, Probs = Ranking, P = 2 * ncol(X),
  Initialized = Initialize(ncol(X), P), mu = 1/ncol(X),
  StopFunction = Stop, IterationsMax, IterationsMin = IterationsMax/2,
  nCores = 1, ...)
}
\arguments{
\item{X}{A n*p matrix of predictors.}

\item{Y}{A n*1 matrix of responses.}

\item{ObjectiveFunction}{An objective criterion/fitness function. Defaults to AIC.}

\item{Probs}{The probability of parents being selected. Defaults to Ranking.}

\item{P}{Population size for generation. Must be an even integer. Defaults to 2p.}

\item{Initialized}{A matrix initialized the population. Defaults to Initialize(p, P).}

\item{mu}{The mutation rate has to be a number between 0 and 1. Defaults to 1/p.}

\item{StopFunction}{A stop criterion. Defaults to Stop function.}

\item{IterationsMax}{The maximum number of iterations.}

\item{IterationsMin}{The minimum number of iterations. Default to half of IterationsMax}

\item{nCores}{Number of cores used for parallelization. Defaults to 1.}
}
\value{
A list with the elements
\item{FinalGeneration}{The generation of the last iteration}
\item{FittestInd}{The fittest individual of the the last iteration}
}
\description{
This function implement a genetic algorithm for variable selection in regression problems.
}
\examples{
X <- matrix(rnorm(30 * 100), ncol = 30)
beta <- 100 * rnorm(10)
select <- sample(1:ncol(X), size = 10, replace = F)
Y <- X[,select] \%*\% matrix(beta, ncol = 1) + rnorm(100)
beta1 <- select(X, Y, IterationsMax = 200)
sum(lm(Y ~X [,select])$residuals^2)
sum(lm(Y ~X [,1:ncol(X)*beta1$FittestInd])$residuals^2)
}
