% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{select Function}
\usage{
select(X, Y, ObjectiveFunction = AIC, Probs = Ranking, P = 2 * ncol(X),
  Initialized = Initialize(ncol(X), P), mu = 1/ncol(X),
  StopFunction = Stop, Iterations, nCores = 1, ...)
}
\arguments{
\item{X}{A n*p matrix of predictors.}

\item{Y}{A n*1 matrix of responses.}

\item{ObjectiveFunction}{An objective criterion/fitness function. Defaults to AIC.}

\item{Probs}{The probability of parents being selected. Defaults to Ranking.}

\item{P}{Population size for generation. Must be an even integer. Defaults to 2p.}

\item{Initialized}{A matrix initialized the population. Defaults to Initialize(p, P).}

\item{mu}{The mutation rate has to be a number between 0 and 1. Defaults to 1/p.}

\item{StopFunction}{A stop criterion. Defaults to Stop function.}

\item{Iterations}{Number of iterations.}

\item{nCores}{Number of cores used for parallelization. Defaults to 1.}
}
\value{
Return the fittest individual in the population.
}
\description{
This function implement a genetic algorithm for variable selection in regression problems.
}
\examples{
select()

}
