% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_parents}
\alias{select_parents}
\title{select_parents}
\usage{
select_parents(fitness_values, mechanism, random = random, P, c)
}
\arguments{
\item{fitness_values}{(vector P): fitness_value of each of the candidate of the current generation}

\item{P}{(int): number of candidates per generation}

\item{mechanism:}{user defined rank-based selection mechanism,
must be one of c("replace_all","tournament", "partial_replace")}

\item{random:}{A boolean value(T/F), "T" if choosing 1 parent selected proportional to fitness + 1 parent
random selected and "F" if 2 parents selected proportional to to fitness}

\item{generation}{gap: proportion of the generation to be replaced by offspring
output:}
}
\value{
parents (matrix P x 2): each row is a pair of indices of parents

candidate(P x c): Each row is a candidate model for breeding
}
\description{
This function returns pairs of parents for breeding
}
