% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Selection function}
\usage{
selection(AUCTable, population, NChrSel)
}
\arguments{
\item{AUCTable}{Table with chromosomes scored.}

\item{population}{Total population of chromosomes}

\item{NChrSel}{Number of chromosomes to select.}
}
\description{
This function allows you to select chromosomes based in scored given by the fitness function.
}
\examples{
Selection(AUCTable,population,NChrSel)
}
\keyword{cats}
