% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-revise.R
\name{selection}
\alias{selection}
\title{\code{selection} returns the parents population and the corresponding fitness values}
\usage{
selection(population, fitnessVec, eliteRate, selectMethod, c = NULL,
  k = NULL)
}
\arguments{
\item{population}{a matrix}

\item{fitnessVec}{a vector of length n}

\item{selectMethod}{a function of selection method, can be gaLRselection-linear selection,
gaExpSelection-nonlinear selection, gaRWselection-Roulette Wheel selection,
gaTNselection-Tournament selection}

\item{c}{a constant specifies the exponential base in nonlinear ranking selection}

\item{k}{number of random selection from population in tournament selection}
}
\description{
\code{selection} returns the parents population and the corresponding fitness values
}
\examples{
numVar <- 6
N <- 50
population <- matrix(rbinom(numVar*N, 1, prob = 0.5), N, numVar)
fitnessVec <- seq(15, 50, length.out=50)
c <- 0.5
k <- 5
eliteRate <- 0.05
selection(population, fitnessVec, selectMethod=gaLRselection, eliteRate)
selection(population, fitnessVec, selectMethod=gaExpSelection, eliteRate, c)
selection(population, fitnessVec, selectMethod=gaRWselection, eliteRate)
selection(population, fitnessVec, selectMethod=gaTNselection, eliteRate, k)

TBD
}
