% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selection}
\alias{selection}
\title{Genetic Algorithms Selection Operation}
\usage{
selection(type, pop_fitness)
}
\arguments{
\item{type}{The type of selection mechanism. Either \code{'oneprop'},
\code{'twoprop'}. See page 76 and 81 of G/H.}

\item{pop_fitness}{A vector of fitness values for the population}
}
\value{
The mutated or unmutated candidate.
}
\description{
Performs a selection operation for the passed type on the
passed fitness values.
}
\examples{
initPopAIC <- unlist(lapply(initPop, function(x) AIC(x$fit)))
selection('oneprop', initPopAIC)
selection('twoprop', initPopAIC)

}
