% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentSelection.R
\name{selection}
\alias{selection}
\title{Selection of highest fitness vectors}
\usage{
selection(AICs, k, population, populationSize)
}
\arguments{
\item{AICs}{fitness values (AICs of regression models)}

\item{k}{parameter used to scale ranking. Must be between 1 & 2}

\item{population}{matrix of population vectors}
}
\value{
a matrix of parent vectors that will be used to produce offsprings
}
\description{
Selects the best fit vectors used to produce offsprings.
Uses roulette wheel to select
}
