% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectparent.R
\name{selectparents}
\alias{selectparents}
\title{select parents in the genetic algorithms}
\usage{
selectparents(originalparents, couplenum, method, fitness, subsetnum = NULL)
}
\arguments{
\item{originalparents}{is a casenum * variablenum matrix}

\item{couplenum}{number of couples selected}

\item{method}{the method to select the parents, "twopropselection", "onepropselection" or "tournament"}

\item{fitness}{goodness of fit ~ a casenum * 1 vector , default is AICvalue}

\item{subsetnum}{In method "tournament", we partition the set of chromosomes in generation t into "subsetnum" disjoint subsets}
}
\value{
Returns the selected parents ~ list
}
\description{
In this selection mechanism, parents are chosen to produce offspring in the next stage.
Three different methods are provided: "twopropselection", "onepropselection" and "tournament"
}
\author{
Lei Zhang
}
