% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectparent1.R
\name{tournament}
\alias{tournament}
\title{Method : tournament selection
the set of parents in generation t is randomly partitioned into k disjoint
subsets of equal size(perhaps with a few remaining parents temporarily ignored)
The best individual in each group is chosen as a parent.
Additional random partitionings are carried out until sufficient parents have
been generated.Parents are then paired randomly for breeding.}
\usage{
tournament(originalparents, couplenum, subsetnum, fitness)
}
\arguments{
\item{originalparents}{: a casenum * variablenum matrix}

\item{couplenum}{: number of couples selected}

\item{fitness:}{goodness of fit ~ a casenum * 1 vector , default is AICvalue}
}
\description{
Returns the selected parents ~ list
}
\author{
Lei Zhang
}
