% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tournament}
\alias{tournament}
\title{Tournament based selection}
\usage{
tournament(fitness, K)
}
\arguments{
\item{fitness}{fitness for individual solutions from fit}

\item{K}{Number of tournaments}
}
\value{
Returns a matrix pop_size by 2 documenting the row index of each parent in population
}
\description{
Select K parents randomly and pick the parent with highest fitness as the first parent
Select the other parent randomly
Iterate this process for pop_size times
}
\examples{
tournament(fitness = c(0.1, 0.14, 0.06, 0.1, 0.12, 0.24, 0.06, 0.03, 0.05, 0.1), K = 4)
}
\author{
Chloe Zhang
}
