% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectparent.R
\name{twopropselection}
\alias{twopropselection}
\title{Method : twopropselection}
\usage{
twopropselection(originalparents, couplenum, fitness)
}
\arguments{
\item{originalparents}{: a casenum * variablenum matrix}

\item{couplenum}{: number of couples selected}

\item{fitness:}{goodness of fit ~ a casenum * 1 vector , default is AICvalue}
}
\description{
Select one parent with probability proportional to fitness(or fitnessrank)
and to select the other parent completely at random
Returns the selected parents ~ list
}
\author{
Lei Zhang
}
